+++ Ready_ant_server/ant_basic/ant_rs.json
@@ -1,1 +1,1 @@
-{"Ant":{"angle":[55821.5999994708,55821.6999994708,55821.7999994708],"pos":[[14.044702142333989,126.37509118058809],[34.17348501016869,156.05065026158783],[54.302833937803285,185.7264346051104]]}}
+{"Ant":{"angle":[12151.20000002602,12151.30000002602,12151.40000002602],"pos":[[195.81151593252707,399.95613669306704],[215.4625331946382,429.94852186247897],[235.1135642786715,459.9402979540176]]}}
