--- Ready_ant_server/static/draw_cansvas/setting_canvas.js
@@ -1,86 +1,0 @@
-const canvas = document.getElementById('canvas');
-canvas.width = 1920
-canvas.height = 900
-const ctx = canvas.getContext('2d');
-canvas.style.backgroundColor = "yellow"
-document.body.appendChild(canvas);
-
-// let line = new Path2D("M -1 -1 h 2 v 2 h -2 Z")
-let among_us_origin = new Path2D(`m0-5c.674.066 1.17.224 1.5.474.284.214.592.566.782.888.304.516.456 1.108.482 1.872l.012.376.062.044c.052.036.098.044.342.056.42.022.814.1 1.104.22.146.062.27.17.332.29.128.252.24.78.294 1.398.062.718.11 2.656.082 3.29-.016.366-.028.456-.07.606-.06.204-.194.48-.338.694-.142.212-.414.48-.564.552-.116.056-.13.058-.39.056-.15-.002-.354-.018-.46-.038l-.19-.032-.052.058c-.08.09-.098.214-.084.596.022.668-.11 1.444-.294 1.734-.052.082-.084.106-.284.202-.266.13-.672.256-1.006.312-.302.05-.334.05-.59-.008-.628-.142-1.132-.492-1.28-.888-.048-.126-.058-.242-.118-1.19-.02-.328-.042-.6-.046-.606-.006-.004-.068-.018-.138-.03-.126-.02-.224-.014-.36.026l-.072.02-.044.322c-.14 1.022-.356 1.632-.608 1.712-.25.078-1.13.102-1.414.036-.4-.09-.666-.348-.872-.848-.198-.476-.302-.986-.486-2.356-.212-1.58-.268-2.594-.172-3.14.04-.23.064-1.078.032-1.15-.012-.028-.072-.114-.132-.19-.142-.18-.276-.452-.332-.68-.124-.504.03-1.14.42-1.73.112-.17.396-.47.69-.73l.158-.14-.034-.054c-.062-.094-.004-.224.226-.51.326-.406.812-.83 1.238-1.08.468-.276.816-.368 1.62-.436.434-.036.674-.036 1.054.002zm-1.088.728c-.814.056-1.24.226-1.648.658-.14.148-.274.388-.274.488l0 .06.124-.012c.636-.064 1.316-.058 1.772.016.352.058.784.152.964.21.512.168.722.494.784 1.222.048.554-.052 1.17-.278 1.718-.31.748-.606 1.08-1.086 1.21-.1.026-.236.038-.58.044-1.01.022-2.29-.138-2.688-.336-.066-.032-.12-.056-.124-.054-.01.012-.11.79-.138 1.074-.084.854.01 1.802.308 3.142.14.636.144.654.14 1.062-.002.312.004.388.036.502.11.372.366.594.732.636.276.032.574.026.654-.016.084-.04.162-.186.258-.48.1-.306.102-.552.01-.786-.028-.066-.046-.078-.214-.146-.378-.148-.68-.372-.774-.57-.026-.056-.042-.106-.036-.112.006-.006.108.004.228.02.542.076.75.094 1.208.104.426.008.53.004.798-.032.34-.046.666-.054.784-.02.042.014.102.05.132.082l.056.058 0 .954 0 .954.056.096c.104.178.412.368.688.424.566.114.92-.004 1.104-.366.1-.198.144-.406.262-1.234.346-2.458.466-4.856.308-6.23-.018-.148-.076-.522-.13-.832-.294-1.688-.558-2.47-.972-2.872-.128-.124-.612-.484-.716-.534-.114-.054-.332-.082-.79-.1-.57-.024-.64-.024-.958-.002zm-1.744 1.918c-.456.102-.872.304-1.268.614-.228.176-.298.258-.378.428-.096.208-.132.356-.146.59-.01.182-.008.21.034.29.152.296.596.574 1.348.838.366.13 1.334.254 1.568.202.324-.072.724-.238.992-.414.198-.13.404-.334.472-.466.03-.06.082-.192.116-.29.18-.524.126-1.086-.13-1.37-.112-.124-.21-.182-.416-.246-.634-.196-1.714-.284-2.192-.176zm6.04 1.794-.13.03.008.684c.012.94-.026 2.422-.108 4.26-.018.404-.028.74-.022.746.006.006.128.012.272.014l.26.002.154-.076c.274-.136.522-.428.614-.722.23-.726.108-3.578-.19-4.49-.108-.334-.214-.442-.466-.478-.12-.018-.216-.01-.392.03z`)
-let among_us = new Path2D(`M 0.1885 -7.9271 C -0.2753 -7.9271 -0.739 -7.9271 -1.2027 -7.9271 C -1.3504 -7.9271 -1.4981 -7.9271 -1.6457 -7.9271 C -1.7268 -7.9271 -1.8021 -7.9486 -1.8451 -7.8673 C -1.9316 -7.7039 -1.8995 -7.4352 -1.8296 -7.2692 C -1.7951 -7.1874 -1.7771 -6.9928 -1.8207 -6.9104 C -1.864 -6.8285 -2.2885 -6.9151 -2.3834 -6.9104 C -2.6176 -6.8986 -2.7356 -6.5556 -2.4831 -6.4762 C -2.358 -6.4369 -2.2408 -6.3732 -2.1154 -6.3322 C -1.9333 -6.2726 -1.7128 -6.2524 -1.5217 -6.2524 C -1.2925 -6.2524 -1.0698 -6.3052 -0.8438 -6.3123 C -0.6719 -6.3176 -0.4803 -6.3418 -0.3144 -6.3876 C -0.1976 -6.4198 -0.0498 -6.4061 0.0733 -6.4275 C 0.2429 -6.4569 0.4348 -6.4319 0.6071 -6.4319 C 0.7947 -6.4319 0.9822 -6.4319 1.1698 -6.4319 C 1.2752 -6.4319 1.4309 -6.4316 1.5154 -6.5116 C 1.5712 -6.5645 1.6234 -6.6299 1.5995 -6.711 C 1.5678 -6.8191 1.2744 -6.8293 1.1853 -6.8306 C 0.992 -6.8334 0.614 -6.7579 0.4631 -6.9148 C 0.3365 -7.0465 0.3836 -7.1989 0.3236 -7.349 C 0.276 -7.4679 0.2682 -7.6004 0.2682 -7.7278 C 0.2682 -7.7578 0.2605 -8.0279 0.184 -7.9515 C 0.1368 -7.9043 0.2231 -7.9119 0.2283 -7.8673 C 0.2324 -7.8329 0.2669 -7.8189 0.2682 -7.7832 C 0.2699 -7.7382 0.2682 -7.693 0.2682 -7.648 C 0.2682 -7.5115 0.2883 -7.4135 0.3435 -7.2892 C 0.354 -7.2656 0.3466 -7.2345 0.3479 -7.2094 C 0.35 -7.1708 0.3858 -7.1464 0.3878 -7.1097 C 0.3906 -7.0602 0.4244 -7.0255 0.4277 -6.9702 C 0.4296 -6.9381 0.6658 -6.8705 0.7112 -6.8705 C 0.8641 -6.8705 1.0169 -6.8705 1.1698 -6.8705 C 1.3387 -6.8705 1.416 -6.8493 1.5442 -6.7664 C 1.8056 -6.5972 1.3502 -6.4951 1.2252 -6.4673 C 0.9744 -6.4116 0.7384 -6.4119 0.4787 -6.4119 C 0.2997 -6.4119 0.1201 -6.392 -0.0552 -6.392 C -0.2207 -6.392 -0.339 -6.2923 -0.5005 -6.2923 C -0.5898 -6.2923 -0.6792 -6.2923 -0.7685 -6.2923 C -0.9428 -6.2923 -1.117 -6.2923 -1.2913 -6.2923 C -1.3826 -6.2923 -1.4203 -6.2048 -1.5217 -6.2724 C -1.5641 -6.3007 -1.6134 -6.3261 -1.6612 -6.2879 C -1.6895 -6.2652 -1.7427 -6.2357 -1.7809 -6.2325 C -1.8273 -6.2286 -1.883 -6.1568 -1.9293 -6.1373 C -2.0613 -6.0814 -2.1583 -5.9411 -2.2328 -5.8293 C -2.3545 -5.6468 -2.4182 -5.4227 -2.4233 -5.2002 C -2.4285 -4.9713 -2.4432 -4.7408 -2.4432 -4.5091 C -2.4432 -4.3924 -2.4432 -4.2758 -2.4432 -4.1591 C -2.4432 -4.0775 -2.2407 -3.957 -2.2881 -3.9043 C -2.3803 -3.8019 -2.3435 -3.6004 -2.3435 -3.4657 C -2.3435 -3.1828 -2.4053 -2.8011 -2.297 -2.542 C -2.2723 -2.4829 -2.366 -2.4919 -2.4033 -2.4556 C -2.4453 -2.4147 -2.4832 -2.3418 -2.4986 -2.285 C -2.5316 -2.1629 -2.5736 -2.0469 -2.6071 -1.9262 C -2.6406 -1.8056 -2.7242 -1.7082 -2.7866 -1.6027 C -2.8643 -1.4712 -2.9572 -1.3414 -3.0524 -1.2217 C -3.1454 -1.1048 -3.223 -0.9758 -3.3293 -0.8695 C -3.4021 -0.7967 -3.4937 -0.7475 -3.5641 -0.6746 C -3.6959 -0.5379 -3.7757 -0.3677 -3.8432 -0.1917 C -3.9402 0.0612 -4.0855 0.2935 -4.1733 0.5504 C -4.2766 0.853 -4.2999 1.1334 -4.2973 1.452 C -4.2952 1.7169 -4.2744 2.0373 -4.1578 2.2805 C -3.8867 2.846 -3.128 3.093 -2.5385 3.0979 C -2.3819 3.0992 -2.2055 3.1185 -2.0533 3.0736 C -1.9472 3.0422 -1.8496 2.974 -1.7454 2.934 C -1.5152 2.8457 -1.2863 2.7799 -1.0919 2.6194 C -0.8895 2.4524 -0.7424 2.2868 -0.6046 2.0612 C -0.4974 1.8858 -0.3967 1.6697 -0.2657 1.5118 C -0.1811 1.41 -0.0795 1.2858 0.029 1.2084 C 0.1237 1.1407 0.2778 1.1425 0.3878 1.1441 C 0.6443 1.1478 0.824 1.3339 1.0258 1.4675 C 1.2463 1.6135 1.4538 1.7901 1.666 1.9482 C 2.0701 2.2494 2.4738 2.5904 2.9752 2.708 C 3.252 2.773 3.5482 2.851 3.8369 2.8144 C 4.0047 2.7931 4.2212 2.7689 4.3508 2.6438 C 4.6617 2.3436 4.734 1.8434 4.734 1.4321 C 4.734 0.9277 4.5928 0.4294 4.3043 0.0121 C 4.0953 -0.2901 3.8203 -0.5249 3.5467 -0.7654 C 3.4268 -0.8708 3.2872 -0.9495 3.1679 -1.0534 C 3.0606 -1.1469 2.9646 -1.2621 2.8644 -1.3635 C 2.6785 -1.5517 2.5301 -1.7651 2.3771 -1.9793 C 2.301 -2.0858 2.2366 -2.1936 2.1688 -2.305 C 2.1101 -2.4016 2.0732 -2.5089 2.0138 -2.604 C 1.8904 -2.8014 1.8236 -3.057 1.7635 -3.2819 C 1.7002 -3.5181 1.6268 -3.7545 1.5486 -3.9863 C 1.465 -4.2341 1.3796 -4.4927 1.265 -4.7284 C 1.1659 -4.9324 1.1002 -5.153 0.966 -5.3354 C 0.8239 -5.5284 0.686 -5.7131 0.5362 -5.9002 C 0.4471 -6.0116 0.3437 -6.1083 0.2438 -6.2081 C 0.1405 -6.3115 -0.0641 -6.4319 -0.2147 -6.4319 C -0.5781 -6.4319 -0.9741 -6.4447 -1.3267 -6.3566 C -1.4191 -6.3335 -1.5103 -6.2978 -1.6059 -6.2923 C -1.677 -6.2883 -1.7376 -6.2363 -1.8052 -6.2325 C -1.9461 -6.2247 -2.0542 -6.0882 -2.1441 -5.9933 C -2.3302 -5.7969 -2.3615 -5.5862 -2.4587 -5.3553 C -2.5064 -5.242 -2.4831 -5.0737 -2.4831 -4.9521 C -2.4831 -4.8087 -2.511 -4.6905 -2.4033 -4.5888 C -2.3769 -4.5639 -2.2962 -4.4799 -2.2638 -4.4781 C -2.2094 -4.475 -2.1271 -4.4473 -2.0799 -4.4183 C -1.9989 -4.3684 -1.761 -4.3714 -1.71 -4.4581 C -1.6842 -4.502 -1.5963 -4.4788 -1.5615 -4.5379 C -1.5446 -4.5667 -1.5076 -4.5489 -1.4907 -4.5777 C -1.4721 -4.6092 -1.4518 -4.6307 -1.4264 -4.6575 C -1.3913 -4.6947 -1.4065 -4.7689 -1.4065 -4.817 C -1.4065 -4.906 -1.4347 -4.9277 -1.5106 -4.9765 C -1.7321 -5.1189 -2.1609 -5.193 -2.4033 -5.0717 C -2.5025 -5.0222 -2.5229 -4.9255 -2.5229 -4.817 C -2.5229 -4.676 -2.4928 -4.6667 -2.3989 -4.5822 C -2.363 -4.5499 -2.3045 -4.555 -2.2638 -4.5335 C -2.2183 -4.5094 -2.155 -4.5008 -2.1043 -4.498 C -1.988 -4.4916 -1.8674 -4.4781 -1.7454 -4.4781 C -1.6304 -4.4781 -1.543 -4.5609 -1.4663 -4.6376 C -1.4215 -4.6823 -1.3959 -4.678 -1.3865 -4.7572 C -1.3781 -4.8289 -1.3666 -4.7942 -1.3666 -4.7372 C -1.3666 -4.513 -1.4663 -4.329 -1.4663 -4.1037 C -1.4663 -3.9904 -1.3609 -3.8573 -1.2271 -3.9243 C -1.1297 -3.973 -1.0622 -4.1192 -0.9479 -4.1192 C -0.8593 -4.1192 -0.7679 -4.1311 -0.6843 -4.0993 C -0.6334 -4.0799 -0.4945 -4.0475 -0.4894 -4.1392 C -0.482 -4.2725 -0.5227 -4.373 -0.5293 -4.498 C -0.5321 -4.5514 -0.6041 -4.6053 -0.6245 -4.6619 C -0.6459 -4.7212 -0.6646 -4.7744 -0.6334 -4.8369 C -0.5495 -5.0046 -0.4047 -5.0189 -0.2435 -5.0762 C -0.0403 -5.1484 0.213 -5.116 0.4277 -5.116 C 0.5925 -5.116 0.647 -4.9174 0.647 -4.7771 C 0.647 -4.4254 -0.0843 -4.4427 -0.3144 -4.5135 C -0.4271 -4.5482 -0.5891 -4.5662 -0.5891 -4.7173 C -0.5891 -4.8481 -0.648 -4.9279 -0.4783 -4.9809 C -0.4242 -4.9978 -0.3742 -5.0529 -0.31 -5.0562 C -0.2231 -5.0608 -0.1361 -5.0762 -0.0508 -5.0762 C 0.1205 -5.0762 0.2918 -5.0762 0.4631 -5.0762 C 0.5535 -5.0762 0.5653 -4.9912 0.5916 -4.9255 C 0.6009 -4.9024 0.6337 -4.7986 0.647 -4.7971 C 0.6761 -4.7936 0.6958 -4.7771 0.7268 -4.7771 C 0.7486 -4.7771 0.694 -4.76 0.6758 -4.7771 C 0.6174 -4.8321 0.7521 -4.9033 0.7821 -4.9322 C 0.853 -5.0003 0.9299 -5.0961 1.0457 -5.0961 C 1.1206 -5.0961 1.1497 -5.0428 1.2052 -5.0363 C 1.2243 -5.0341 1.232 -5.0019 1.2207 -5.0119 C 1.1999 -5.0305 1.1706 -5.042 1.1499 -5.0607 C 1.1268 -5.0814 1.0324 -5.0909 1.0059 -5.0673 C 0.9851 -5.0488 0.9 -4.9582 0.8862 -4.9566 C 0.8555 -4.9529 0.7837 -4.8828 0.7622 -4.8613 C 0.7116 -4.8108 0.7219 -4.7971 0.6426 -4.7971 C 0.5909 -4.7971 0.5544 -4.8313 0.5074 -4.8524 C 0.4386 -4.8834 0.3614 -4.8768 0.2881 -4.8768 C 0.2075 -4.8768 0.1331 -4.8967 0.0533 -4.8967 C 0.0233 -4.8967 -0.0996 -4.8768 -0.0663 -4.8768 C 0.3246 -4.8768 -0.1972 -4.8834 -0.1062 -4.8015 C -0.0851 -4.7825 0.1622 -4.7797 0.184 -4.8015 C 0.2221 -4.8395 -0.0404 -4.905 -0.0508 -4.817 C -0.0571 -4.763 0.1849 -4.7983 0.0954 -4.8569 C 0.071 -4.8728 -0.1493 -4.9036 -0.1062 -4.8369 C -0.0834 -4.8017 0.0969 -4.7875 0.0888 -4.8569 C 0.0794 -4.9363 -0.074 -4.88 -0.1062 -4.8569 C -0.195 -4.7929 0.0694 -4.7829 0.1087 -4.8015 C 0.1619 -4.8267 0.165 -4.8768 0.2328 -4.8768 C 0.3894 -4.8768 0.5116 -4.8931 0.647 -4.8214 C 0.779 -4.7515 0.6274 -4.5652 0.5318 -4.5578 C 0.485 -4.5542 0.4487 -4.5193 0.3989 -4.5179 C 0.3666 -4.517 0.3189 -4.5272 0.2881 -4.5135 C 0.2326 -4.4888 0.1694 -4.498 0.1087 -4.498 C -0.0089 -4.498 -0.1339 -4.4837 -0.2457 -4.5224 C -0.2864 -4.5365 -0.441 -4.5494 -0.4495 -4.5977 C -0.4576 -4.6433 -0.5256 -4.6631 -0.5448 -4.6176 C -0.5881 -4.5147 -0.5492 -4.3162 -0.5492 -4.199 C -0.5492 -4.1314 -0.5403 -4.0892 -0.6135 -4.0638 C -0.6535 -4.05 -0.6857 -4.0382 -0.7286 -4.0395 C -0.7749 -4.0408 -0.7984 -4.0776 -0.8438 -4.0793 C -0.9292 -4.0826 -1.007 -4.0864 -0.8771 -4.0239 C -0.8218 -3.9974 -0.5768 -3.9962 -0.7242 -4.0594 C -0.764 -4.0765 -1.0004 -4.0892 -0.8638 -4.0638 C -0.6954 -4.0325 -0.8977 -4.0594 -0.9634 -4.0594 C -1.0966 -4.0594 -1.1868 -4.0619 -1.2514 -3.9398 C -1.2609 -3.9219 -1.2946 -3.9037 -1.3134 -3.8999 C -1.3663 -3.8893 -1.2056 -3.8979 -1.1517 -3.8999 C -1.106 -3.9016 -1.2634 -3.9796 -1.3068 -3.9796 C -1.339 -3.9796 -1.3666 -3.986 -1.3666 -3.9442 C -1.3666 -3.8796 -1.2333 -3.9579 -1.2138 -3.9708 C -1.1258 -4.0288 -1.3094 -4.0939 -1.3622 -4.0594 C -1.5015 -3.9683 -1.2647 -3.9703 -1.2647 -3.9952 C -1.2647 -4.0221 -1.4439 -3.9722 -1.4663 -3.9841 C -1.552 -4.0295 -1.4862 -4.2317 -1.4862 -4.2986 C -1.4862 -4.4078 -1.4264 -4.4909 -1.4264 -4.5977 C -1.4264 -4.8603 -1.8914 -4.817 -2.0799 -4.817 C -2.1997 -4.817 -2.0858 -4.7936 -2.0245 -4.7971 C -2.0083 -4.798 -2.1221 -4.9059 -2.1951 -4.8369 C -2.2761 -4.7605 -2.1359 -4.6951 -2.0843 -4.6819 C -1.9867 -4.6568 -1.9152 -4.7745 -2.0046 -4.7771 C -2.1504 -4.7814 -2.1401 -4.7093 -2.0046 -4.7173 C -1.9698 -4.7194 -1.9435 -4.7701 -1.9736 -4.7971 C -2.0114 -4.8309 -2.0969 -4.8195 -2.1441 -4.817 C -2.1924 -4.8145 -2.1429 -4.7666 -2.1198 -4.7616 C -2.0774 -4.7525 -1.9993 -4.7445 -1.9647 -4.7771 C -1.8711 -4.8656 -2.2022 -4.7618 -1.9891 -4.8015 C -1.872 -4.8233 -1.7135 -4.9126 -1.6059 -4.817 C -1.5276 -4.7475 -1.3055 -4.7118 -1.3965 -4.5479 C -1.4727 -4.4107 -1.8088 -4.4581 -1.9448 -4.4581 C -2.0587 -4.4581 -2.2638 -4.5857 -2.2638 -4.4072 C -2.2638 -4.2634 -2.32 -4.3384 -2.215 -4.3983 C -2.16 -4.4298 -2.0908 -4.4183 -2.0245 -4.4183 C -1.9476 -4.4183 -1.8729 -4.3983 -1.7942 -4.3983 C -1.7652 -4.3983 -1.6376 -4.4328 -1.6258 -4.4061 C -1.5937 -4.3338 -1.6059 -4.2462 -1.6059 -4.169 C -1.6059 -4.0744 -1.4984 -4.0868 -1.4874 -4.0483 C -1.4767 -4.0109 -1.409 -3.9589 -1.4774 -3.9198 C -1.5096 -3.9014 -1.5668 -3.8425 -1.5859 -3.8091 C -1.6211 -3.7476 -1.6891 -3.7635 -1.7244 -3.7105 C -1.7409 -3.6856 -1.7739 -3.6894 -1.7853 -3.6695 C -1.8003 -3.6432 -1.8068 -3.6164 -1.834 -3.6008 C -1.9319 -3.5449 -1.6136 -3.4613 -1.5549 -3.4613 C -1.3384 -3.4613 -1.1266 -3.4619 -0.9169 -3.5001 C -0.7723 -3.5264 -0.607 -3.541 -0.4595 -3.541 C -0.3664 -3.541 -0.2734 -3.541 -0.1804 -3.541 C -0.1577 -3.541 -0.0634 -3.5551 -0.0508 -3.5709 C 0.0027 -3.6379 0.2237 -3.5421 0.1874 -3.6695 C 0.1614 -3.7603 -0.0848 -3.808 -0.1604 -3.8501 C -0.2079 -3.8764 -0.2448 -3.9087 -0.3 -3.921 C -0.3089 -3.9229 -0.4485 -4.0088 -0.4495 -4.0107 C -0.4617 -4.032 -0.6026 -4.0868 -0.6079 -4.0683 C -0.6326 -3.9819 -0.581 -3.9418 -0.5104 -3.911 C -0.3889 -3.8578 -0.2714 -3.7888 -0.1505 -3.7404 C -0.1255 -3.7304 -0.0114 -3.6508 -0.063 -3.6706 C -0.1738 -3.7133 -0.2549 -3.7906 -0.3598 -3.839 C -0.3996 -3.8574 -0.5093 -3.9321 -0.5093 -3.9863 C -0.5093 -4.103 -0.4401 -3.7385 -0.3388 -3.6806 C -0.2822 -3.6483 -0.3276 -3.6786 -0.362 -3.7205 C -0.4209 -3.7921 -0.5091 -3.8541 -0.5293 -3.9497 C -0.5523 -4.059 -0.5638 -3.8909 -0.5581 -3.8567 C -0.5409 -3.7539 -0.5288 -3.6605 -0.4706 -3.5732 C -0.4469 -3.5376 -0.4497 -3.6765 -0.4517 -3.6895 C -0.4649 -3.7716 -0.5093 -3.8091 -0.5093 -3.7005 C -0.5093 -3.4971 -0.5725 -3.5908 -0.6289 -3.7094 C -0.6906 -3.8388 -0.7286 -3.5497 -0.7286 -3.6507 C -0.7286 -3.7994 -0.8483 -3.5787 -0.8483 -3.6064 C -0.8483 -3.6303 -0.8323 -3.8001 -0.8704 -3.7947 C -0.9328 -3.7858 -0.9887 -3.6439 -1.0067 -3.5931 C -1.0494 -3.4721 -0.9663 -3.8107 -1.0798 -3.7072 C -1.095 -3.6932 -1.2246 -3.5689 -1.226 -3.5765 C -1.2297 -3.597 -1.2101 -3.7232 -1.2204 -3.7138 C -1.251 -3.686 -1.3085 -3.5831 -1.3555 -3.5831 C -1.3844 -3.5831 -1.3449 -3.6457 -1.3921 -3.6363 C -1.467 -3.6213 -1.5203 -3.548 -1.5937 -3.5333 C -1.5942 -3.5332 -1.5501 -3.624 -1.5472 -3.6385 C -1.536 -3.6944 -1.6173 -3.5467 -1.6635 -3.5133 C -1.6849 -3.4979 -1.7397 -3.4602 -1.7144 -3.5145 C -1.6906 -3.5655 -1.6457 -3.63 -1.6457 -3.6872 C -1.6457 -3.724 -1.7399 -3.5762 -1.71 -3.5975 C -1.6327 -3.6527 -1.6454 -3.6941 -1.6225 -3.5909 C -1.6113 -3.5405 -1.5749 -3.4711 -1.5161 -3.4835 C -1.4543 -3.4964 -1.4257 -3.5816 -1.3777 -3.5953 C -1.3588 -3.6007 -1.362 -3.4889 -1.3146 -3.5145 C -1.2873 -3.5292 -1.2266 -3.6181 -1.1971 -3.5887 C -1.1629 -3.5544 -1.1531 -3.5122 -1.0986 -3.5023 C -0.9438 -3.4741 -1.0139 -3.731 -0.897 -3.6197 C -0.8067 -3.5337 -0.7096 -3.5052 -0.6877 -3.6585 C -0.6748 -3.7485 -0.4568 -3.5738 -0.4196 -3.582 C -0.3257 -3.6029 -0.4162 -3.7005 -0.3077 -3.7005 C -0.2843 -3.7005 -0.0978 -3.6856 -0.0918 -3.7127 C -0.0747 -3.7896 -0.2674 -3.8972 -0.321 -3.9508 C -0.4178 -4.0476 -0.5492 -4.1427 -0.5492 -4.2887 C -0.5492 -4.3727 -0.5854 -4.5578 -0.4573 -4.5578 C -0.2524 -4.5578 -0.0475 -4.5578 0.1574 -4.5578 C 0.2717 -4.5578 0.3783 -4.5609 0.4864 -4.5789 C 0.6067 -4.5989 0.6691 -4.7493 0.7766 -4.7971 C 0.8792 -4.8426 0.8987 -4.9915 1.0158 -5.0175 C 1.1273 -5.0422 1.1244 -5.0381 1.1654 -4.9255 C 1.1922 -4.8517 1.2294 -4.8064 1.244 -4.7262 C 1.2685 -4.5916 1.3497 -4.4757 1.3858 -4.3496 C 1.398 -4.3067 1.4445 -4.2852 1.4445 -4.2289 C 1.4445 -4.1675 1.4517 -4.1194 1.4633 -4.0671 C 1.48 -3.992 1.469 -3.9145 1.4954 -3.839 C 1.5402 -3.7111 1.5837 -3.5875 1.6328 -3.4613 C 1.6838 -3.33 1.7333 -3.1982 1.7945 -3.0703 C 1.8778 -2.8961 1.902 -2.6985 1.9828 -2.5254 C 2.0691 -2.3404 2.1243 -2.158 2.2486 -1.9882 C 2.3704 -1.8218 2.5183 -1.674 2.6939 -1.5651 C 2.7705 -1.5175 2.8131 -1.4825 2.86 -1.4089 C 2.96 -1.2517 3.0729 -1.0999 3.189 -0.9581 C 3.25 -0.8835 3.2889 -0.7938 3.3484 -0.7211 C 3.3745 -0.6893 3.4473 -0.6134 3.457 -0.5793 C 3.486 -0.4777 3.6234 -0.456 3.7029 -0.401 C 3.7534 -0.366 3.8141 -0.3234 3.8668 -0.2825 C 3.9821 -0.1928 4.1401 -0.2029 4.2644 -0.1319 C 4.3365 -0.0907 3.9236 0.1407 3.879 0.155 C 3.7905 0.1835 3.7881 0.1618 3.8003 0.0642 C 3.823 -0.1172 3.9661 -0.1886 4.1049 -0.2803 C 4.1652 -0.32 4.3155 -0.3728 4.3353 -0.4519 C 4.3555 -0.5328 4.403 -0.6054 4.4128 -0.689 C 4.4191 -0.7419 4.435 -0.8037 4.435 -0.8562 C 4.435 -0.9483 4.4224 -0.672 4.4051 -0.5815 C 4.3754 -0.4267 4.3482 -0.1877 4.2046 -0.092 C 4.1079 -0.0275 4.0106 0.0133 3.9155 0.0786 C 3.8645 0.1137 3.7088 0.2258 3.7394 0.2946 C 3.7513 0.3213 3.8788 0.3089 3.9089 0.3189 C 4.0072 0.3517 4.1031 0.3856 4.2024 0.4164 C 4.2928 0.4445 4.3897 0.4573 4.476 0.4962 C 4.509 0.511 4.5605 0.539 4.5956 0.5449 C 4.6395 0.5522 4.6974 0.5507 4.7407 0.5449 C 4.8127 0.5353 4.6256 0.5449 4.6122 0.5438 C 4.4678 0.5322 4.3209 0.4715 4.1802 0.4364 C 4.1208 0.4215 4.0399 0.4062 3.9864 0.3765 C 3.9284 0.3443 3.8651 0.3218 3.8003 0.3079 C 3.6565 0.277 3.7606 0.3019 3.848 0.3167 C 3.9659 0.3368 4.0812 0.3865 4.2013 0.3865 C 4.253 0.3865 4.294 0.4264 4.3453 0.4264 C 4.3887 0.4264 4.4785 0.5025 4.497 0.5349 C 4.5369 0.6047 4.688 0.642 4.7518 0.6845 C 4.7953 0.7135 4.8055 0.7401 4.8127 0.7941 C 4.821 0.8566 4.8624 0.8996 4.8714 0.9625 C 4.8858 1.0635 4.8906 1.1598 4.9223 1.2549 C 4.9501 1.3381 4.9496 1.4274 4.8404 1.4199 C 4.5231 1.398 4.2312 1.0502 3.972 0.8894 C 3.9197 0.8569 3.8967 0.8544 3.8967 0.9182 C 3.8967 1.0643 3.9977 1.1134 4.1094 1.1951 C 4.2665 1.31 4.4511 1.3951 4.5989 1.5218 C 4.6632 1.5769 4.7175 1.6849 4.7651 1.7555 C 4.782 1.7807 4.9733 2.0986 4.9733 2.0313 C 4.9733 1.7623 4.7638 1.7361 4.5568 1.6326 C 4.4481 1.5782 4.3985 1.4704 4.3231 1.3811 C 4.1999 1.2352 4.1008 1.1116 3.9177 1.0544 C 3.8799 1.0426 3.8576 1.0838 3.9067 1.0544 C 3.9248 1.0435 3.9703 1.0206 3.9454 0.9957 C 3.9096 0.9599 3.7059 1.0794 3.6707 1.0921 C 3.5529 1.1345 3.313 1.243 3.2997 1.3955 C 3.2843 1.5725 3.498 1.6282 3.498 1.7821 C 3.498 2.0704 3.6085 2.3087 3.6973 2.5807 C 3.7759 2.8214 3.8635 3.0809 3.8757 3.3361 C 3.8764 3.3521 3.8888 3.5205 3.8956 3.4834 C 3.934 3.272 3.9359 3.0566 3.8181 2.8709 C 3.6801 2.6533 3.6709 2.3875 3.5389 2.1642 C 3.4608 2.032 3.3784 1.8689 3.3784 1.7123 C 3.3784 1.6778 3.3845 1.6371 3.3761 1.6038 C 3.3724 1.5888 3.34 1.5592 3.3562 1.5916 C 3.3917 1.6625 3.4518 1.7263 3.509 1.7799 C 3.638 1.9006 3.7875 1.9867 3.9056 2.1199 C 4.0321 2.2626 4.1233 2.3896 4.2046 2.5641 C 4.2768 2.719 4.3647 2.861 4.5303 2.9251 C 4.5829 2.9455 4.8047 2.9999 4.692 2.8853 C 4.5451 2.7359 4.3503 2.6724 4.2046 2.5308 C 4.0498 2.3803 3.9446 2.1557 3.817 1.9815 C 3.7357 1.8706 3.6969 1.7829 3.5777 1.7023 C 3.5129 1.6585 3.3542 1.5346 3.2654 1.5561 C 3.0628 1.6053 2.9349 1.7227 2.9641 1.9416 C 2.9984 2.1988 3.1354 2.4162 2.9796 2.666 C 2.8017 2.9512 2.4854 3.1421 2.2464 3.3726 C 2.187 3.4298 2.2714 3.3977 2.3372 3.397 C 2.4153 3.3961 2.4548 3.398 2.4812 3.3239 C 2.5243 3.2026 2.5387 3.1122 2.634 3.0137 C 2.7075 2.9379 2.7703 2.8731 2.8201 2.7789 C 2.967 2.5014 3.0671 2.2027 3.1103 1.8884 C 3.1217 1.8052 3.1548 1.664 3.0616 1.6182 C 2.9571 1.5668 2.8438 1.5398 2.7359 1.4985 C 2.6427 1.4629 2.5625 1.3965 2.4657 1.3679 C 2.4321 1.3579 2.418 1.7683 2.4125 1.8042 C 2.3441 2.2501 2.0494 2.6197 1.7834 2.9694 C 1.7504 3.0128 1.6701 3.1112 1.604 3.0979 C 1.5679 3.0907 1.685 2.9688 1.6948 2.9584 C 1.7311 2.92 1.7605 2.8715 1.7635 2.8188 C 1.7675 2.7456 1.8832 2.6809 1.934 2.6482 C 2.1498 2.5095 2.2448 2.2809 2.2663 2.0324 C 2.2828 1.8418 2.3018 1.6546 2.3018 1.4631 C 2.3018 1.4356 2.3287 1.3971 2.304 1.3789 C 2.2264 1.3219 2.1656 1.3409 2.1046 1.4011 C 1.7861 1.7152 1.4129 1.9308 1.0701 2.2118 C 0.9997 2.2696 0.9623 2.3003 0.9106 2.3802 C 0.9021 2.3933 0.7918 2.5246 0.791 2.522 C 0.7349 2.3443 1.1697 2.2153 1.2805 2.1454 C 1.4105 2.0634 1.4082 1.9198 1.4888 1.802 C 1.5354 1.7339 1.5561 1.6594 1.6239 1.6071 C 1.6738 1.5686 1.7042 1.5284 1.7435 1.4786 C 1.779 1.4337 1.9014 1.3079 1.903 1.2549 C 1.9056 1.1687 1.8816 1.0444 1.7701 1.0444 C 1.3785 1.0444 1.1265 1.3847 0.9173 1.6713 C 0.8207 1.8036 0.7185 1.9865 0.5916 2.0922 C 0.4996 2.1689 0.6199 1.9907 0.6426 1.9704 C 0.8186 1.8127 1.0297 1.7878 1.0258 1.514 C 1.0234 1.3467 1.0059 1.1798 1.0059 1.0112 C 1.0059 0.8665 1.0026 0.7848 0.884 0.6834 C 0.7999 0.6113 0.7154 0.5359 0.616 0.4862 C 0.5643 0.4603 0.5888 0.607 0.5872 0.6501 C 0.5841 0.7338 0.5535 0.8116 0.5562 0.8982 C 0.5655 1.203 0.7696 1.4881 0.8464 1.7799 C 0.9101 2.0221 1.024 2.4119 0.8575 2.6394 C 0.8327 2.6732 0.7423 2.7843 0.7467 2.6859 C 0.7533 2.5382 0.8262 2.3895 0.8464 2.2406 C 0.8742 2.0359 0.9047 1.8366 0.9062 1.6292 C 0.908 1.3693 0.8061 1.151 0.7821 0.896 C 0.7661 0.7255 0.6337 0.4322 0.4233 0.546 C 0.3348 0.5939 0.2048 0.5845 0.1264 0.6346 C 0.0114 0.7081 0.2424 1.204 0.2726 1.2814 C 0.34 1.4537 0.4131 1.6227 0.4897 1.791 C 0.5212 1.8601 0.5541 1.9295 0.5628 2.0058 C 0.5754 2.1161 0.5207 2.2022 0.4565 2.2872 C 0.4039 2.3568 0.4546 2.4255 0.5119 2.4799 C 0.566 2.5313 0.6599 2.7159 0.6669 2.6416 C 0.6973 2.3231 0.5778 2.0519 0.4521 1.771 C 0.388 1.6279 0.3162 1.4944 0.2638 1.3457 C 0.2065 1.1833 0.1858 1.0134 0.1286 0.8517 C 0.0901 0.7428 0.0774 0.6258 -0.0508 0.6258 C -0.0594 0.6258 0.0632 0.5666 0.0866 0.5526 C 0.2595 0.4495 0.4138 0.3272 0.4631 0.1229 C 0.5025 -0.04 0.4812 -0.1899 0.4321 -0.3467 C 0.4163 -0.3974 0.3757 -0.4899 0.4521 -0.5062 C 0.6349 -0.5454 0.5012 -0.275 0.4631 -0.1917 C 0.3872 -0.0251 0.2734 0.1636 0.2682 0.3466 C 0.2659 0.4263 0.183 0.5876 0.1087 0.6213 C 0.0714 0.6383 -0.0409 0.7082 -0.0751 0.7011 C -0.1171 0.6924 -0.1454 0.6656 -0.1903 0.6656 C -0.3104 0.6656 -0.4967 0.8271 -0.6002 0.8849 C -0.6591 0.9178 -0.7468 0.9666 -0.7885 1.0156 C -0.861 1.101 -0.6783 1.2098 -0.6312 1.2549 C -0.5404 1.3417 -0.5036 1.4715 -0.4495 1.5827 C -0.3711 1.7439 -0.3254 1.9244 -0.3144 2.1033 C -0.3007 2.3261 -0.3983 2.5375 -0.4207 2.7568 C -0.4388 2.9344 -0.3361 2.992 -0.2236 3.1046 C -0.1486 3.1795 -0.1914 3.0518 -0.2081 2.996 C -0.2906 2.7193 -0.3645 2.4753 -0.3941 2.1875 C -0.4081 2.0521 -0.4269 1.9068 -0.5337 1.8109 C -0.6375 1.7176 -0.8052 1.6902 -0.8948 1.5827 C -0.9155 1.5578 -0.9297 1.5154 -0.928 1.483 C -0.9268 1.4591 -0.8641 1.3765 -0.8771 1.3679 C -0.9279 1.334 -1.1114 1.3819 -1.1429 1.4277 C -1.1938 1.5018 -1.2048 1.6953 -1.2115 1.7843 C -1.2277 1.9957 -1.2271 2.206 -1.2271 2.4179 C -1.2271 2.6112 -1.2104 2.8126 -1.2315 3.0049 C -1.2496 3.1696 -1.3326 3.3043 -1.4065 3.4479 C -1.4429 3.5187 -1.4691 3.5654 -1.3932 3.6207 C -1.3698 3.6378 -1.2271 3.74 -1.2271 3.6827 C -1.2271 3.5289 -1.2335 3.4005 -1.1916 3.2486 C -1.1507 3.1001 -1.1141 2.9463 -1.0875 2.7944 C -1.0283 2.4568 -1.1673 2.1113 -1.1673 1.771 C -1.1673 1.7089 -1.1385 1.5716 -1.1872 1.5229 C -1.2014 1.5087 -1.3066 1.4513 -1.3179 1.4609 C -1.3742 1.5087 -1.4027 1.6561 -1.4353 1.7223 C -1.4791 1.8112 -1.5001 1.822 -1.5948 1.822 C -1.6812 1.822 -1.8495 1.7898 -1.9293 1.8419 C -1.9996 1.8879 -1.9536 1.9872 -2.0046 2.0413 C -2.1279 2.1718 -2.2285 2.3742 -2.2638 2.5485 C -2.3029 2.7418 -2.2785 2.9067 -2.173 3.0758 C -2.0962 3.1987 -1.939 3.3968 -2.0489 3.5476 C -2.1232 3.6495 -2.2607 3.6601 -2.3745 3.6717 C -2.489 3.6833 -2.4138 3.6087 -2.359 3.5764 C -2.1406 3.4475 -1.9904 3.2667 -2.0733 3.0004 C -2.1287 2.8225 -2.238 2.6272 -2.2438 2.44 C -2.2481 2.3039 -2.1883 2.1694 -2.1287 2.0501 C -2.0712 1.9352 -2.2017 1.9826 -2.2992 1.977 C -2.4209 1.9701 -2.5618 1.967 -2.6802 1.9992 C -2.8807 2.0536 -3.041 2.2009 -3.1764 2.3514 C -3.2132 2.3922 -3.3553 2.4636 -3.3005 2.4599 C -3.2645 2.4575 -3.2417 2.3833 -3.2407 2.4422 C -3.2379 2.6105 -3.2469 2.7792 -3.2363 2.9473 C -3.2276 3.0832 -3.1885 3.2728 -3.2717 3.3948 C -3.3293 3.4792 -3.5749 3.6679 -3.5065 3.5919 C -3.4407 3.5188 -3.3229 3.5112 -3.3049 3.397 C -3.287 3.2828 -3.3418 3.1738 -3.3758 3.0691 C -3.4141 2.9511 -3.4145 2.7885 -3.3913 2.666 C -3.3758 2.5842 -3.307 2.485 -3.3204 2.4001 C -3.3339 2.3144 -3.5684 2.2748 -3.6151 2.3448 C -3.7001 2.4724 -3.7548 2.6507 -3.7856 2.7989 C -3.8151 2.9407 -3.7695 3.0969 -3.8122 3.2353 C -3.8482 3.3519 -3.9357 3.4331 -4.0514 3.4701 C -4.0876 3.4816 -4.2641 3.5274 -4.2929 3.4679 C -4.3236 3.4046 -4.0413 3.297 -4.0071 3.2818 C -3.7167 3.1521 -3.6728 2.7994 -3.6793 2.5109 C -3.6818 2.3974 -3.7192 2.2952 -3.7192 2.1808 C -3.7192 2.0785 -3.7277 1.9847 -3.8587 1.9815 C -4.1293 1.9747 -4.2425 2.3402 -4.5122 2.296 C -4.5979 2.282 -4.5964 2.2014 -4.5964 2.1321 C -4.5964 2.0347 -4.6745 2.0612 -4.7559 2.0612 C -4.8862 2.0612 -4.7309 2.0931 -4.6806 2.121 C -4.5737 2.1804 -4.4885 2.2194 -4.3638 2.2207 C -4.1993 2.2224 -4.1354 2.183 -4.0027 2.09 C -3.9684 2.0659 -3.8815 2.0126 -3.8787 1.9615 C -3.8754 1.9037 -3.9047 1.8575 -3.9141 1.802 C -3.9268 1.7272 -3.9133 1.6503 -3.9673 1.5872 C -4.0889 1.4447 -4.2942 1.3659 -4.4457 1.2593 C -4.5263 1.2026 -4.5947 1.1184 -4.6894 1.0843 C -4.7741 1.0538 -4.9571 1.0732 -4.9818 0.958 C -4.9918 0.9115 -4.8794 0.882 -4.86 0.9137 C -4.8198 0.9795 -4.8596 1.0432 -4.7714 1.0954 C -4.6194 1.1852 -4.4489 1.2194 -4.3062 1.3302 C -4.226 1.3924 -4.1283 1.4305 -4.0559 1.503 C -3.9776 1.5812 -3.9727 1.6525 -3.9495 1.7511 C -3.9222 1.8672 -3.9066 1.8852 -3.7878 1.8973 C -3.6936 1.9068 -3.6351 1.8679 -3.6394 1.7688 C -3.6493 1.5408 -3.8164 1.356 -3.8809 1.1441 C -3.9261 0.9954 -3.9362 0.8325 -3.8831 0.6856 C -3.8553 0.6086 -3.7951 0.5264 -3.7989 0.4419 C -3.8043 0.3224 -4.1246 0.2626 -4.2132 0.227 C -4.2478 0.2131 -4.5603 0.0892 -4.4679 0.0875 C -4.4146 0.0864 -4.2776 0.0606 -4.2375 0.1074 C -4.1765 0.1786 -4.207 0.3914 -4.1821 0.484 C -4.1533 0.5917 -4.0911 0.6791 -3.9983 0.7409 C -3.8271 0.855 -3.6866 0.9883 -3.6793 1.2084 C -3.6775 1.263 -3.6593 1.3139 -3.6593 1.3679 C -3.6593 1.4312 -3.7093 1.4672 -3.7435 1.5185 C -3.786 1.5822 -3.6788 1.3371 -3.6682 1.308 C -3.6098 1.1487 -3.5754 0.9811 -3.5198 0.8207 C -3.4727 0.685 -3.437 0.5389 -3.4245 0.3954 C -3.4062 0.1847 -3.5376 0.037 -3.666 -0.1141 C -3.7619 -0.2271 -3.7636 -0.3146 -3.728 -0.453 C -3.7162 -0.499 -3.6984 -0.5961 -3.6416 -0.6059 C -3.5582 -0.6203 -3.3963 -0.6232 -3.3293 -0.5616 C -3.2567 -0.4949 -3.3102 -0.3049 -3.3293 -0.2293 C -3.4019 0.0591 -3.4135 0.3296 -3.3847 0.6258 C -3.379 0.6835 -3.381 0.7738 -3.4046 0.8296 C -3.4216 0.8697 -3.5492 0.9362 -3.491 0.9979 C -3.4361 1.056 -3.3013 0.7995 -3.2916 0.7808 C -3.2284 0.6584 -3.218 0.5131 -3.2008 0.3776 C -3.1731 0.1594 -3.1363 -0.0719 -3.1897 -0.2891 C -3.2386 -0.4876 -3.3493 -0.7834 -3.5154 -0.9094 C -3.5441 -0.9312 -3.5723 -0.9306 -3.6084 -0.9337 C -3.7506 -0.9461 -3.8923 -0.9582 -4.0337 -0.978 C -4.1571 -0.9954 -4.4354 -0.9869 -4.5321 -1.0932 C -4.6095 -1.1783 -4.9502 -0.9301 -4.8666 -1.0091 C -4.7552 -1.1143 -4.4605 -1.1942 -4.3372 -1.0777 C -4.1552 -0.9059 -3.9158 -0.908 -3.6726 -0.885 C -3.5998 -0.8781 -3.5307 -0.8718 -3.4578 -0.8695 C -3.397 -0.8676 -3.3261 -0.8641 -3.2783 -0.8208 C -3.2484 -0.7936 -3.2538 -0.7395 -3.2052 -0.7388 C -3.0806 -0.737 -2.8936 -0.8069 -2.7755 -0.8496 C -2.7073 -0.8742 -2.5336 -0.9534 -2.5229 -1.0334 C -2.509 -1.138 -2.5673 -1.1284 -2.6514 -1.1531 C -2.8508 -1.2116 -3.0494 -1.2722 -3.2495 -1.328 C -3.4792 -1.3921 -3.7246 -1.4054 -3.9495 -1.4831 C -4.0281 -1.5103 -4.0831 -1.5434 -4.1223 -1.616 C -4.1612 -1.688 -4.2098 -1.7235 -4.2929 -1.7268 C -4.4193 -1.7318 -4.0396 -1.7188 -3.9141 -1.7024 C -3.8375 -1.6924 -3.6593 -1.6871 -3.6593 -1.5784 C -3.6593 -1.5298 -3.6704 -1.4654 -3.6549 -1.4189 C -3.6167 -1.3042 -3.3447 -1.3002 -3.2518 -1.2793 C -3.0778 -1.2403 -2.9359 -1.2272 -2.8619 -1.0556 C -2.8323 -0.9869 -2.632 -1.1128 -2.6869 -1.0091 C -2.7136 -0.9586 -2.8542 -0.9348 -2.8619 -0.8695 C -2.8675 -0.8213 -2.9862 -0.7375 -3.0169 -0.7056 C -3.0927 -0.627 -3.121 -0.459 -3.1366 -0.3534 C -3.1494 -0.2664 -3.136 -0.1197 -3.1764 -0.0432 C -3.1975 -0.0034 -3.2328 0.0456 -3.2407 0.0875 C -3.2498 0.1363 -3.376 0.0703 -3.3957 0.0587 C -3.5471 -0.0302 -3.6392 -0.1878 -3.7546 -0.3135 C -3.904 -0.4762 -4.1761 -0.359 -4.3682 -0.3722 C -4.4954 -0.381 -4.6969 -0.481 -4.7847 -0.3556 C -4.7966 -0.3386 -4.7714 -0.3827 -4.7581 -0.391 C -4.6945 -0.4305 -4.6115 -0.4465 -4.5388 -0.4586 C -4.3966 -0.4823 -4.3365 -0.4488 -4.2287 -0.3633 C -4.0235 -0.2006 -3.7423 -0.3322 -3.5597 -0.1496 C -3.4 0.0101 -3.2649 0.2184 -3.193 0.4341 C -3.1194 0.6551 -3.4404 0.7374 -3.5674 0.8362 C -3.6091 0.8686 -3.6254 0.9306 -3.6704 0.947 C -3.7464 0.9746 -3.6944 1.006 -3.7723 0.9436 C -3.9909 0.7687 -4.2197 0.7454 -4.4912 0.7454 C -4.5868 0.7454 -4.6567 0.7125 -4.7404 0.6678 C -4.7717 0.6511 -4.769 0.7019 -4.7382 0.7044 C -4.6789 0.7091 -4.5299 0.6736 -4.4956 0.7365 C -4.4657 0.7912 -4.3509 0.8103 -4.3018 0.8251 C -4.1954 0.8572 -4.0882 0.8841 -3.9839 0.9226 C -3.8765 0.9621 -3.7612 1.0099 -3.6494 1.0322 C -3.5789 1.0463 -3.4474 1.0337 -3.501 1.1364 C -3.586 1.2993 -3.6485 1.4692 -3.738 1.6315 C -3.7995 1.7429 -3.9334 1.8836 -3.8853 2.0235 C -3.8075 2.2499 -3.5466 2.427 -3.8343 2.6305 C -3.9684 2.7253 -4.1308 2.7809 -4.2752 2.8587 C -4.3021 2.8732 -4.4386 2.925 -4.4446 2.9672 C -4.448 2.9905 -4.2611 3.016 -4.2408 3.016 C -4.1729 3.016 -4.1091 2.9435 -4.0669 2.8963 C -3.9767 2.7952 -3.9098 2.6879 -3.8731 2.5552 C -3.8596 2.5064 -3.8604 2.3903 -3.8078 2.3603 C -3.6795 2.287 -3.4413 2.4004 -3.3193 2.4389 C -3.21 2.4734 -3.0558 2.4278 -2.9582 2.491 C -2.879 2.5423 -2.9217 2.8582 -2.9217 2.9318 C -2.9217 3.1397 -2.9332 3.3222 -2.7999 3.4955 C -2.728 3.589 -2.5941 3.6487 -2.5351 3.747 C -2.5117 3.786 -2.5155 3.8553 -2.5529 3.8843 C -2.5962 3.9181 -2.6171 3.8326 -2.6326 3.8068 C -2.6672 3.7491 -2.6747 3.6814 -2.6824 3.6152 C -2.6971 3.4903 -2.7414 3.4085 -2.8198 3.3106 C -2.962 3.1328 -2.9295 2.9059 -2.9117 2.6892 C -2.8958 2.4947 -2.9424 2.0346 -2.6714 2.0036 C -2.4652 1.98 -2.3031 2.0068 -2.1032 2.0402 C -1.946 2.0664 -1.765 2.0235 -1.649 1.9117 C -1.5656 1.8312 -1.5015 1.6951 -1.4043 1.6326 C -1.3048 1.5686 -1.1818 1.5827 -1.0687 1.5827 C -0.9944 1.5827 -0.7853 1.5412 -0.7286 1.606 C -0.6384 1.7091 -0.5885 1.8403 -0.4883 1.9405 C -0.3965 2.0323 -0.273 2.1001 -0.2014 2.2107 C -0.0549 2.4371 0.0329 2.6877 0.1486 2.9273 C 0.2247 3.085 0.1763 2.9344 0.1364 2.8576 C 0.0688 2.7274 -0.0039 2.6077 -0.0608 2.4721 C -0.1061 2.364 -0.1337 2.2171 -0.2047 2.1254 C -0.2965 2.0071 -0.4824 1.9846 -0.6068 1.9106 C -0.7565 1.8215 -0.933 1.7138 -0.9945 1.5462 C -1.0307 1.4473 -0.9867 1.4022 -0.9269 1.3324 C -0.9045 1.3063 -0.8682 1.3122 -0.8682 1.2737 C -0.8682 1.238 -0.9175 1.1991 -0.9402 1.1729 C -1.0326 1.0663 -1.2117 1.0343 -1.3422 1.0057 C -1.5562 0.9588 -1.7223 0.9157 -1.834 1.1264 C -1.8943 1.2401 -1.9363 1.4421 -2.0334 1.5307 C -2.149 1.6361 -2.3501 1.6359 -2.4964 1.6625 C -2.5508 1.6724 -2.5723 1.825 -2.5938 1.8696 C -2.6177 1.919 -2.7428 2.0293 -2.7124 1.9837 C -2.6172 1.8409 -2.4208 1.8001 -2.297 1.6913 C -2.1752 1.5842 -2.2303 1.4696 -2.2006 1.3313 C -2.1827 1.2476 -1.9731 1.2019 -1.9093 1.1452 C -1.8861 1.1245 -1.83 1.0599 -1.8606 1.0655 C -2.1121 1.1112 -2.2672 1.2197 -2.3424 1.4642 C -2.3499 1.4887 -2.5012 2.0108 -2.5905 1.9394 C -2.5931 1.9373 -2.1262 1.638 -2.1021 1.6193 C -1.9488 1.5005 -1.8237 1.336 -1.8063 1.1364 C -1.7974 1.0335 -1.8623 1.1349 -1.8949 1.1474 C -1.9291 1.1606 -2.1215 0.9178 -2.1475 0.896 C -2.2369 0.821 -2.3913 0.6854 -2.513 0.6667 C -2.6415 0.647 -2.7932 0.7998 -2.884 0.865 C -3.0627 0.9933 -3.2638 1.0611 -3.2573 1.3091 C -3.2523 1.4998 -3.1489 1.6883 -3.1654 1.8773 C -3.177 2.0113 -3.2407 1.965 -3.2407 1.8607 C -3.2407 1.7246 -3.1654 1.6035 -3.1222 1.4764 C -3.0879 1.3759 -3.0285 1.2424 -3.0967 1.1441 C -3.1527 1.0635 -3.2509 0.965 -3.2097 1.1186 C -3.1505 1.3392 -3.1609 1.5614 -3.1609 1.7876 C -3.1609 2.0605 -3.2066 1.1604 -2.997 0.9857 C -2.8693 0.8793 -2.6694 0.9449 -2.5828 0.8063 C -2.527 0.7171 -2.5348 0.6848 -2.616 0.6258 C -2.7773 0.5084 -2.8899 0.3462 -3.0092 0.1871 C -3.1365 0.0174 -3.0767 -0.1485 -3.1022 -0.3401 C -3.1143 -0.4304 -3.2694 -0.3824 -3.3326 -0.3921 C -3.5572 -0.4267 -3.418 -0.612 -3.3592 -0.7532 C -3.2319 -1.0587 -3.1747 -1.4702 -3.2728 -1.7955 C -3.2986 -1.8811 -3.3802 -1.9853 -3.3802 -2.0757 C -3.3802 -2.1002 -3.4712 -2.1593 -3.4644 -2.1632 C -3.2856 -2.2654 -3.4249 -1.8271 -3.429 -1.8087 C -3.466 -1.6396 -3.4421 -1.4404 -3.398 -1.2749 C -3.3879 -1.2372 -3.3254 -1.0933 -3.2706 -1.1298 C -3.24 -1.1502 -3.2274 -1.2017 -3.1997 -1.2295 C -3.1322 -1.2969 -3.0754 -1.3721 -3.0192 -1.4488 C -2.9485 -1.5451 -2.8505 -1.6175 -2.7733 -1.708 C -2.6676 -1.8318 -2.5628 -1.9827 -2.5628 -2.1532 C -2.5628 -2.2187 -2.5327 -2.2865 -2.5041 -2.3437 C -2.4666 -2.4188 -2.4595 -2.5046 -2.4321 -2.583 C -2.3648 -2.7753 -2.3435 -2.9748 -2.3435 -3.18 C -2.3435 -3.3879 -2.3524 -3.5919 -2.3646 -3.7991 C -2.3736 -3.9534 -2.3787 -4.126 -2.4421 -4.2687 C -2.4844 -4.3639 -2.473 -4.4936 -2.4919 -4.5966 C -2.5078 -4.6827 -2.5074 -4.7718 -2.4931 -4.858 C -2.4673 -5.0122 -2.4575 -5.1612 -2.4233 -5.3154 C -2.3974 -5.4318 -2.318 -5.533 -2.2837 -5.6444 C -2.2644 -5.7071 -2.208 -5.7472 -2.1652 -5.7939 C -2.1247 -5.8381 -2.126 -5.8868 -2.0932 -5.9346 C -2.0277 -6.0299 -2.0149 -6.1199 -1.926 -6.2015 C -1.8465 -6.2744 -1.9981 -6.3156 -2.0445 -6.3322 C -2.1648 -6.3752 -2.2968 -6.3807 -2.4122 -6.4419 C -2.4989 -6.4877 -2.499 -6.5455 -2.5329 -6.6202 C -2.6154 -6.8017 -2.2893 -6.7567 -2.215 -6.6611 C -2.1928 -6.6325 -2.1886 -6.5316 -2.2339 -6.5316 C -2.3627 -6.5316 -2.4454 -6.5538 -2.4908 -6.68 C -2.5675 -6.8929 -2.1405 -6.869 -2.1264 -6.721 C -2.1138 -6.588 -2.1422 -6.4286 -2.3025 -6.4153 C -2.3892 -6.408 -2.5769 -6.6042 -2.4299 -6.6257 C -2.4074 -6.629 -2.3225 -6.607 -2.3059 -6.6257 C -2.286 -6.6481 -2.2842 -6.6845 -2.2627 -6.7088 C -2.2118 -6.7661 -2.1092 -6.7646 -2.0533 -6.7176 C -1.9824 -6.658 -1.9036 -6.5389 -1.968 -6.4463 C -2.0371 -6.3469 -2.1973 -6.5143 -2.1973 -6.5936 C -2.1973 -6.7205 -2.1018 -6.797 -1.9847 -6.7896 C -1.8461 -6.781 -1.7926 -6.611 -1.7864 -6.4928 C -1.7776 -6.3257 -1.935 -6.2986 -2.0733 -6.3643 C -2.2361 -6.4417 -2.0141 -6.5768 -1.9415 -6.598 C -1.8107 -6.6362 -1.6139 -6.6356 -1.5881 -6.4618 C -1.567 -6.3193 -1.7259 -6.3226 -1.8241 -6.3566 C -2.303 -6.5224 -1.6379 -7.2554 -1.4519 -6.7719 C -1.4165 -6.68 -1.3899 -6.49 -1.5383 -6.4972 C -1.7386 -6.507 -1.758 -6.7652 -1.69 -6.9093 C -1.6675 -6.957 -1.5973 -7.0191 -1.5383 -7.0001 C -1.4619 -6.9754 -1.453 -6.8377 -1.453 -6.7752 C -1.453 -6.506 -1.7749 -6.6636 -1.8484 -6.8107 C -1.9262 -6.9663 -1.859 -7.1769 -1.6967 -7.2449 C -1.5275 -7.3157 -1.4608 -7.1068 -1.4674 -6.9757 C -1.4704 -6.9154 -1.4894 -6.9038 -1.5427 -6.9403 C -1.6401 -7.0068 -1.6992 -7.1101 -1.7166 -7.226 C -1.7369 -7.3615 -1.676 -7.5816 -1.504 -7.5816 C -1.4126 -7.5816 -1.2026 -7.3778 -1.3544 -7.339 C -1.453 -7.3138 -1.6391 -7.3668 -1.6922 -7.4586 C -1.7389 -7.5393 -1.6876 -7.7161 -1.5948 -7.7477 C -1.4643 -7.7921 -1.3274 -7.4824 -1.453 -7.4354 C -1.5816 -7.3871 -1.8681 -7.536 -1.7786 -7.6945 C -1.6812 -7.8672 -1.281 -7.808 -1.144 -7.7344 C -1.0821 -7.7012 -1.0639 -7.6244 -1.0842 -7.5605 C -1.1154 -7.462 -1.2408 -7.4545 -1.3267 -7.4564 C -1.662 -7.4639 -1.6418 -7.7731 -1.3489 -7.8651 C -1.1731 -7.9203 -0.9723 -7.8944 -0.9723 -7.6801 C -0.9723 -7.484 -1.2333 -7.5399 -1.3013 -7.6691 C -1.3909 -7.8394 -1.1717 -7.9015 -1.0476 -7.9061 C -0.8988 -7.9116 -0.671 -7.8464 -0.671 -7.6613 C -0.671 -7.5347 -0.9598 -7.5858 -1.0144 -7.6425 C -1.1328 -7.7654 -0.6021 -7.8057 -0.5171 -7.7654 C -0.3616 -7.6918 -0.5136 -7.516 -0.6323 -7.5118 C -0.7116 -7.509 -0.8192 -7.5149 -0.8793 -7.5749 C -0.9443 -7.6399 -0.9204 -7.7346 -0.8394 -7.771 C -0.7273 -7.8213 -0.3161 -7.9098 -0.2623 -7.7366 C -0.2213 -7.6045 -0.3354 -7.5062 -0.4672 -7.5362 C -0.6759 -7.5836 -0.5796 -7.7715 -0.4285 -7.8119 C -0.3208 -7.8408 -0.0571 -7.8322 -0.094 -7.6547 C -0.1293 -7.4843 -0.4918 -7.4516 -0.4141 -7.607 C -0.3409 -7.7535 -0.0137 -7.7726 0.0755 -7.6414 C 0.1805 -7.487 0.061 -7.2802 -0.125 -7.3457 C -0.3447 -7.4229 -0.1046 -7.5809 0.0301 -7.5605 C 0.2003 -7.5347 0.3485 -7.2965 0.1663 -7.1751 C 0.0642 -7.107 -0.5002 -7.1156 -0.4617 -7.308 C -0.4328 -7.4525 -0.0169 -7.419 0.019 -7.2947 C 0.0502 -7.1868 -0.1255 -7.1287 -0.1959 -7.1175 C -0.3481 -7.0933 -0.5185 -7.126 -0.6356 -7.2316 C -0.8333 -7.4098 -0.3201 -7.4566 -0.2302 -7.3202 C -0.0706 -7.078 -0.445 -7.0343 -0.588 -7.0854 C -0.7738 -7.1517 -0.9101 -7.4208 -0.6478 -7.4841 C -0.4994 -7.5199 -0.3214 -7.4067 -0.4196 -7.2349 C -0.5441 -7.0171 -1.0298 -7.172 -1.1551 -7.3102 C -1.3463 -7.5213 -1.0109 -7.5689 -0.8582 -7.5273 C -0.7579 -7.4999 -0.6397 -7.4154 -0.6223 -7.3047 C -0.6012 -7.1701 -0.847 -7.1714 -0.918 -7.1729 C -1.0353 -7.1753 -1.3099 -7.1964 -1.3666 -7.3346 C -1.3964 -7.4073 -1.2408 -7.4541 -1.2005 -7.4642 C -1.0337 -7.5059 -0.7694 -7.4763 -0.8737 -7.2415 C -0.97 -7.0249 -1.3547 -7.1296 -1.4962 -7.2382 C -1.5595 -7.2868 -1.4967 -7.3504 -1.4475 -7.3756 C -1.3311 -7.4349 -1.1981 -7.3984 -1.0997 -7.3202 C -0.8152 -7.0941 -1.1926 -6.9365 -1.4021 -6.9569 C -1.5877 -6.975 -1.7303 -7.2077 -1.4729 -7.2338 C -1.3139 -7.2499 -1.1868 -7.1901 -1.1927 -7.0156 C -1.2024 -6.7285 -1.6976 -6.8231 -1.6468 -6.948 C -1.5889 -7.0905 -1.1775 -7.0083 -1.1706 -6.8483 C -1.1614 -6.637 -1.3814 -6.5993 -1.5549 -6.6246 C -1.712 -6.6475 -1.6518 -6.7747 -1.5062 -6.7697 C -1.2567 -6.7611 -1.0929 -6.5179 -1.3644 -6.3798 C -1.4432 -6.3398 -1.8648 -6.3387 -1.7255 -6.5116 C -1.6349 -6.6241 -1.1964 -6.7141 -1.1008 -6.5803 C -0.9813 -6.413 -1.3186 -6.4604 -1.3622 -6.5847 C -1.404 -6.7039 -1.2246 -6.7823 -1.1429 -6.8018 C -1.0177 -6.8318 -0.8701 -6.8249 -0.7696 -6.7354 C -0.5618 -6.5502 -0.8368 -6.4298 -1.0089 -6.4718 C -1.1052 -6.4952 -1.2503 -6.57 -1.2703 -6.68 C -1.2942 -6.8118 -1.1014 -6.8668 -1.01 -6.876 C -0.909 -6.8862 -0.6037 -6.8335 -0.8084 -6.6866 C -0.8814 -6.6342 -1.1261 -6.6153 -1.051 -6.7786 C -0.9721 -6.9499 -0.3527 -7.0454 -0.4684 -6.7697 C -0.5655 -6.5381 -1.0565 -6.6182 -0.876 -6.8074 C -0.7561 -6.9329 -0.1764 -6.904 -0.3698 -6.6977 C -0.4259 -6.6378 -0.5792 -6.6156 -0.5259 -6.7486 C -0.4795 -6.8648 -0.3153 -6.8659 -0.2258 -6.8063 C -0.1693 -6.7686 -0.1606 -6.6869 -0.1992 -6.6335 C -0.2595 -6.55 -0.3892 -6.5435 -0.4816 -6.5471 C -0.7281 -6.5565 -0.5873 -6.7264 -0.4252 -6.7442 C -0.3104 -6.7568 -0.0753 -6.7654 0.0135 -6.6722 C 0.0756 -6.607 -0.0143 -6.471 -0.0807 -6.4474 C -0.1591 -6.4195 -0.2247 -6.5086 -0.2413 -6.5725 C -0.2712 -6.6878 -0.178 -6.7268 -0.0818 -6.7464 C 0.0841 -6.7804 0.4501 -6.7922 0.3912 -6.5172 C 0.3545 -6.346 0.039 -6.4331 0.1087 -6.5725 C 0.1619 -6.6789 0.3836 -6.7455 0.493 -6.7188 C 0.6033 -6.6917 0.4253 -6.5778 0.4033 -6.5703 C 0.1712 -6.4918 0.173 -6.7552 0.3635 -6.8051 C 0.5091 -6.8434 0.7223 -6.8185 0.7223 -6.6312 C 0.7223 -6.4481 0.45 -6.4709 0.6082 -6.5792 C 0.6883 -6.634 1.0705 -6.7166 1.0579 -6.5216 C 1.0538 -6.4573 0.978 -6.4194 0.9239 -6.4075 C 0.7991 -6.3801 0.7471 -6.4328 0.74 -6.5615 C 0.7282 -6.7776 0.7556 -7.0692 0.7866 -7.2714 C 0.8175 -7.4734 0.8555 -7.7432 0.9837 -7.9227 C 1.0472 -8.0116 1.2553 -8.0044 1.2063 -7.8684 C 1.169 -7.7648 1.0158 -7.7599 0.9461 -7.8297 C 0.9031 -7.8726 0.9824 -7.9884 1.0103 -7.8275 C 1.0844 -7.4004 1.0457 -6.9412 1.0457 -6.5094 C 1.0457 -6.4701 1.0349 -6.389 1.0557 -6.3433 C 1.0699 -6.3121 1.0917 -6.0978 1.0612 -6.2081 C 1.0311 -6.3174 0.7389 -6.7 1.0745 -6.618 C 1.1633 -6.5963 1.2616 -6.4563 1.1875 -6.3665 C 1.0906 -6.2492 0.9813 -6.3927 0.987 -6.4906 C 0.994 -6.6093 1.1806 -6.6454 1.2639 -6.5781 C 1.3242 -6.5294 1.3876 -6.3753 1.2783 -6.3388 C 1.2226 -6.3203 0.9591 -6.3247 0.9693 -6.4219 C 0.9791 -6.5149 1.2 -6.4669 1.2506 -6.4496 C 1.4236 -6.3905 1.3356 -6.1138 1.1554 -6.1539 C 1.0951 -6.1673 1.0371 -6.2556 1.0147 -6.3034 C 0.9256 -6.4944 1.1292 -6.5617 1.2938 -6.5504 C 1.4014 -6.543 1.4431 -6.4007 1.337 -6.3477 C 1.1622 -6.2603 0.9475 -6.4011 0.8818 -6.5615 C 0.8647 -6.6033 0.8227 -6.7057 0.8818 -6.7365 C 0.9191 -6.7559 0.9967 -6.759 1.0313 -6.732 C 1.0863 -6.6893 1.0784 -6.5869 1.0358 -6.5393 C 0.973 -6.4691 0.8388 -6.471 0.7544 -6.4773 C 0.6222 -6.4871 0.2108 -6.6187 0.2317 -6.8063 C 0.238 -6.8634 0.349 -6.7972 0.2826 -6.7332 C 0.1722 -6.6266 -0.2147 -6.7057 -0.0896 -6.9015 C -0.06 -6.9478 0.0321 -6.9621 0.081 -6.9458 C 0.1331 -6.9284 0.1002 -6.7944 0.0932 -6.763 C 0.0412 -6.5309 -0.2904 -6.6021 -0.4307 -6.7143 C -0.6181 -6.8642 -0.4736 -6.9857 -0.2911 -6.9857 C -0.1519 -6.9857 -0.0061 -6.9089 -0.0552 -6.7431 C -0.1022 -6.5846 -0.4383 -6.7084 -0.5104 -6.7763 C -0.6924 -6.9476 -0.4458 -7.1069 -0.2723 -7.0798 C -0.0875 -7.051 -0.0125 -6.714 -0.1992 -6.6379 C -0.3438 -6.579 -0.7786 -6.6152 -0.8195 -6.8096 C -0.8656 -7.0289 -0.4291 -6.9643 -0.3343 -6.9259 C -0.1858 -6.8656 -0.1498 -6.6667 -0.3122 -6.598 C -0.3746 -6.5716 -0.5093 -6.5415 -0.5093 -6.6423 C -0.5093 -6.718 -0.4894 -6.7679 -0.4894 -6.8417 C -0.4894 -6.8868 -0.5223 -6.9573 -0.5071 -6.999 C -0.4731 -7.0926 -0.1819 -6.9896 -0.135 -6.9646 C 0.015 -6.885 0.1065 -6.7132 0.1065 -6.546 C 0.1065 -6.3513 -0.1924 -6.4833 -0.2557 -6.5748 C -0.3939 -6.7743 -0.1878 -6.9865 0.0002 -7.0134 C 0.1054 -7.0284 0.2222 -7.0278 0.3136 -6.9669 C 0.3585 -6.9369 0.4441 -6.8559 0.3878 -6.7963 C 0.2326 -6.6319 -0.1339 -6.8478 -0.1937 -7.02 C -0.2668 -7.2309 -0.0106 -7.3124 0.1519 -7.3058 C 0.3787 -7.2965 0.18 -7.011 0.0722 -6.9624 C -0.1426 -6.8656 -0.555 -6.9829 -0.6467 -7.2105 C -0.7475 -7.4609 -0.4352 -7.5041 -0.2645 -7.4398 C 0.0171 -7.3337 -0.1194 -6.9463 -0.3609 -6.8882 C -0.4783 -6.86 -0.9522 -6.9065 -0.9236 -7.0998 C -0.8985 -7.2691 -0.5422 -7.2395 -0.4473 -7.174 C -0.3831 -7.1296 -0.4233 -7.0316 -0.4717 -6.9945 C -0.5609 -6.9262 -0.7057 -6.9099 -0.8139 -6.9059 C -0.9432 -6.9011 -1.2614 -7.0338 -1.01 -7.1529 C -0.8621 -7.223 -0.5639 -7.2047 -0.66 -6.9713 C -0.7582 -6.7328 -1.156 -6.7523 -1.2957 -6.9425 C -1.4266 -7.1206 -1.0862 -7.2444 -0.9845 -7.1197 C -0.9128 -7.0318 -0.9506 -6.8771 -1 -6.7896 C -1.071 -6.6639 -1.1852 -6.6335 -1.3201 -6.6335 C -1.5257 -6.6335 -1.5097 -6.7586 -1.319 -6.7863 C -1.1091 -6.8168 -1.1211 -6.6138 -1.2946 -6.5493 C -1.4796 -6.4805 -1.735 -6.493 -1.9071 -6.5914 C -2.1149 -6.7101 -1.7908 -6.7144 -1.7122 -6.7077 C -1.6619 -6.7034 -1.557 -6.6916 -1.5737 -6.6135 C -1.5882 -6.5462 -1.7052 -6.5468 -1.7543 -6.5449 C -1.8148 -6.5425 -2.2323 -6.5521 -2.1297 -6.7099 C -2.0762 -6.7924 -1.7535 -6.7725 -1.7698 -6.6501 C -1.792 -6.4835 -2.1032 -6.5565 -2.1973 -6.5947 C -2.2112 -6.6004 -2.3508 -6.6615 -2.3336 -6.6988 C -2.3113 -6.747 -2.0781 -6.7368 -2.0489 -6.7376 C -1.9201 -6.7408 -1.7439 -6.7517 -1.6291 -6.6778 C -1.5341 -6.6165 -1.6342 -6.5651 -1.7 -6.5593 C -1.7766 -6.5524 -1.9086 -6.5716 -1.8584 -6.6822 C -1.7989 -6.8131 -1.6153 -6.8311 -1.4929 -6.8472 C -1.3428 -6.8671 -1.1921 -6.8705 -1.041 -6.8705 C -1.0222 -6.8705 -0.9213 -6.8605 -1.0077 -6.8738 C -1.1596 -6.8972 -1.3728 -7.0012 -1.4021 -7.1695 C -1.4405 -7.3905 -1.0655 -7.4069 -0.9269 -7.3445 C -0.8547 -7.3121 -0.945 -7.2384 -0.9779 -7.2238 C -1.0897 -7.1744 -1.2342 -7.1914 -1.35 -7.2138 C -1.4375 -7.2308 -1.7246 -7.3315 -1.5837 -7.4686 C -1.5233 -7.5274 -1.1259 -7.682 -1.093 -7.5339 C -1.0542 -7.3591 -1.4929 -7.3091 -1.597 -7.3612 C -1.6438 -7.3845 -1.6176 -7.4463 -1.5859 -7.4697 C -1.433 -7.5827 -1.1691 -7.5785 -0.9911 -7.5805 C -0.858 -7.5819 -0.533 -7.6037 -0.4351 -7.4741 C -0.3618 -7.377 -0.5543 -7.3122 -0.6146 -7.3113 C -0.7025 -7.31 -0.7806 -7.3654 -0.8006 -7.452 C -0.8367 -7.6083 -0.6362 -7.6628 -0.5237 -7.6868 C -0.3643 -7.7207 -0.1088 -7.7532 0.0389 -7.6558 C 0.1143 -7.606 0.0485 -7.5148 -0.0076 -7.4841 C -0.1427 -7.4102 -0.5463 -7.4016 -0.4916 -7.6502 C -0.444 -7.8664 0.0698 -7.9371 0.1907 -7.7621 C 0.3504 -7.5308 0.0412 -7.2781 -0.1715 -7.2781 C -0.365 -7.2781 -0.4211 -7.4743 -0.1726 -7.4498 C 0.0596 -7.4268 0.3574 -7.2291 0.2472 -6.9558 C 0.2151 -6.8763 0.1257 -6.818 0.0389 -6.8262 C 0.0047 -6.8294 -0.0114 -6.8653 0.0157 -6.8893 C 0.1297 -6.9907 0.5252 -6.9067 0.6027 -6.7819 C 0.6958 -6.6319 0.3169 -6.6232 0.2438 -6.6822 C 0.178 -6.7354 0.3506 -6.7482 0.3579 -6.7486 C 0.5146 -6.7576 0.6768 -6.7402 0.8242 -6.6844 C 0.834 -6.6807 1.0786 -6.579 0.9926 -6.5183 C 0.9821 -6.5109 0.9183 -6.5064 0.9361 -6.5349 C 1.0092 -6.6519 1.2685 -6.61 1.3669 -6.5593 C 1.4314 -6.5261 1.4031 -6.4648 1.3404 -6.4496 C 1.165 -6.407 0.9511 -6.4665 0.7866 -6.5249 C 0.6573 -6.5709 0.5369 -6.655 0.4022 -6.6844 C 0.3079 -6.705 0.2001 -6.6831 0.1065 -6.67 C 0.0081 -6.6563 -0.1206 -6.6162 -0.2203 -6.6368 C -0.2596 -6.6449 -0.1394 -6.7017 -0.1206 -6.7032 C -0.0788 -6.7067 -0.2518 -6.6599 -0.2723 -6.6611 C -0.3489 -6.6657 -0.6123 -6.6937 -0.6068 -6.8184 C -0.6031 -6.9021 -0.4566 -6.9364 -0.3952 -6.9458 C -0.3899 -6.9466 -0.2467 -6.9585 -0.3044 -6.9104 C -0.4465 -6.792 -0.7521 -6.9255 -0.8505 -7.0477 C -0.9377 -7.1559 -0.803 -7.2072 -0.7098 -7.2072 C -0.5985 -7.2072 -0.5491 -7.0735 -0.6755 -7.0433 C -0.7443 -7.0268 -1.1943 -7.0098 -1.0919 -7.1762 C -1.0191 -7.2945 -0.7459 -7.2964 -0.6445 -7.2194 C -0.5972 -7.1835 -0.5985 -7.1359 -0.6622 -7.1308 C -0.845 -7.1161 -1.1229 -7.2063 -1.1229 -7.4309 C -1.1229 -7.5925 -0.9383 -7.5917 -0.9059 -7.4442 C -0.8408 -7.1484 -1.4246 -7.2174 -1.5682 -7.2925 C -1.8288 -7.4288 -1.6246 -7.5712 -1.4198 -7.5816 C -1.3217 -7.5865 -1.0161 -7.4998 -1.2149 -7.3733 C -1.4906 -7.1979 -1.9309 -7.6289 -1.6048 -7.7344 C -1.5433 -7.7543 -1.2433 -7.7269 -1.3865 -7.6314 C -1.4547 -7.586 -1.8529 -7.5458 -1.813 -7.7123 C -1.7979 -7.7751 -1.619 -7.7389 -1.5848 -7.7333 C -1.4398 -7.7098 -1.3418 -7.6412 -1.3002 -7.4941 C -1.2397 -7.2803 -1.3646 -7.0521 -1.5981 -7.0521 C -1.609 -7.0521 -1.6696 -7.0601 -1.628 -7.0798 C -1.5507 -7.1164 -1.4361 -7.0879 -1.3611 -7.0632 C -1.1192 -6.9835 -1.1736 -6.7493 -1.422 -6.7808 C -1.496 -6.7901 -1.6874 -6.841 -1.7044 -6.9347 C -1.7185 -7.0122 -1.4654 -7.005 -1.4386 -7.0045 C -1.2887 -7.0017 -1.0922 -6.9824 -0.9889 -6.8572 C -0.8937 -6.7419 -1.0107 -6.656 -1.113 -6.618 C -1.2308 -6.5741 -1.3694 -6.5608 -1.494 -6.577 C -1.5101 -6.5791 -1.7385 -6.6282 -1.6934 -6.6866 C -1.5873 -6.8239 -0.9046 -6.6665 -1.1495 -6.5183 C -1.3901 -6.3727 -1.8452 -6.3982 -2.0954 -6.5094 C -2.2886 -6.5953 -2.0076 -6.5781 -1.9437 -6.5493 C -1.898 -6.5287 -1.9069 -6.487 -1.9448 -6.4629 C -2.0186 -6.4159 -2.1269 -6.4145 -2.2084 -6.4385 C -2.2821 -6.4603 -2.3461 -6.5151 -2.3679 -6.5903 C -2.3997 -6.7001 -2.2482 -6.7089 -2.1774 -6.7154 C -1.9412 -6.7372 -1.7019 -6.7056 -1.4674 -6.6811 C -1.2755 -6.661 -1.0734 -6.6553 -0.8848 -6.6113 C -0.8391 -6.6006 -0.7847 -6.5767 -0.8627 -6.5615 C -0.9886 -6.5368 -1.1235 -6.5322 -1.2514 -6.5349 C -1.3085 -6.5361 -1.5046 -6.5307 -1.5184 -6.6213 C -1.5343 -6.7266 -1.2088 -6.7066 -1.1761 -6.7066 C -1.0619 -6.7066 -0.6973 -6.7394 -0.6112 -6.6368 C -0.5489 -6.5625 -1.077 -6.5295 -1.0975 -6.5305 C -1.1796 -6.5343 -1.4057 -6.5743 -1.3312 -6.6091 C -1.164 -6.6871 -0.8366 -6.6318 -0.6666 -6.629 C -0.5936 -6.6279 -0.3443 -6.6261 -0.2435 -6.6091 C -0.2324 -6.6072 -0.2051 -6.5972 -0.2147 -6.5914 C -0.2402 -6.576 -0.2727 -6.5775 -0.3022 -6.5737 C -0.5339 -6.5436 -0.7713 -6.5515 -1.0044 -6.5515 C -1.0592 -6.5515 -1.2227 -6.555 -1.1684 -6.5615 C -0.9271 -6.5902 -0.6724 -6.5722 -0.4307 -6.5703 C -0.3392 -6.5696 -0.1041 -6.5689 0.0101 -6.5604 C 0.0937 -6.5541 0.0468 -6.547 0.0024 -6.5426 C -0.1991 -6.523 -0.4047 -6.5316 -0.6068 -6.5316 C -0.6342 -6.5316 -0.7996 -6.531 -0.846 -6.5327 C -0.9209 -6.5354 -0.8308 -6.5407 -0.8173 -6.5415 C -0.5896 -6.5558 -0.3607 -6.5515 -0.1327 -6.5515 C 0.0171 -6.5515 0.4592 -6.564 0.3158 -6.5205 C 0.0779 -6.4483 -0.2129 -6.4759 -0.4584 -6.4729 C -0.5635 -6.4716 -0.5914 -6.4718 -0.6799 -6.4718 C -0.7186 -6.4718 -0.6846 -6.4716 -0.6245 -6.474 C -0.2915 -6.487 0.0413 -6.5034 0.3745 -6.5094 C 0.5201 -6.512 0.5749 -6.5116 0.7046 -6.5116 C 0.769 -6.5116 0.7308 -6.5114 0.6714 -6.5127 C 0.4877 -6.5167 0.3005 -6.5137 0.1176 -6.5316 C -0.1313 -6.5558 0.6174 -6.5515 0.8674 -6.5515 C 0.8833 -6.5515 0.9307 -6.5523 0.915 -6.5493 C 0.8735 -6.5414 0.8315 -6.5367 0.7899 -6.5294 C 0.6994 -6.5133 0.4419 -6.4973 0.5241 -6.4563 C 0.6043 -6.4161 1.3286 -6.495 1.2961 -6.5803 C 1.2672 -6.6559 1.0061 -6.6379 0.9593 -6.6401 C 0.8576 -6.645 0.6135 -6.6942 0.5163 -6.6379 C 0.447 -6.5978 0.6657 -6.5717 0.7456 -6.5659 C 0.913 -6.5537 1.1365 -6.5355 1.2961 -6.6013 C 1.3427 -6.6206 1.2068 -6.6557 1.1565 -6.66 C 1.0052 -6.673 0.8516 -6.6718 0.7013 -6.649 C 0.6792 -6.6456 0.4962 -6.6087 0.6271 -6.5836 C 0.7384 -6.5623 0.8622 -6.5712 0.9737 -6.5847 C 1.0832 -6.598 1.1946 -6.6237 1.2894 -6.6822 C 1.3399 -6.7133 1.3774 -6.7618 1.4312 -6.7874 C 1.464 -6.803 1.4646 -6.8111 1.419 -6.8207 C 1.261 -6.8537 1.0898 -6.8525 0.9294 -6.8594 C 0.8 -6.865 0.494 -6.9232 0.3778 -6.8384 C 0.3546 -6.8214 0.4218 -6.7974 0.4499 -6.7907 C 0.5559 -6.7657 0.6898 -6.753 0.7943 -6.7453 C 0.9944 -6.7307 1.5914 -6.7244 1.3957 -6.7686 C 1.1116 -6.8328 0.7792 -6.7907 0.4897 -6.7907 C 0.4402 -6.7907 0.1513 -6.7964 0.4022 -6.7797 C 0.6272 -6.7647 0.8524 -6.756 1.0779 -6.752 C 1.1486 -6.7507 1.204 -6.7509 1.1244 -6.7509 C 1.0048 -6.7509 0.8851 -6.7509 0.7655 -6.7509 C 0.5319 -6.7509 0.2731 -6.7815 0.0411 -6.7476 C -0.1039 -6.7263 0.3273 -6.6803 0.4731 -6.6645 C 0.7335 -6.6362 1.0098 -6.613 1.2717 -6.6346 C 1.4178 -6.6466 0.9829 -6.6914 0.8364 -6.6977 C 0.5259 -6.7111 0.1566 -6.7571 -0.1505 -6.6888 C -0.1698 -6.6845 -0.1224 -6.6566 -0.1028 -6.6534 C -0.0076 -6.6379 0.1082 -6.6355 0.2017 -6.6324 C 0.3036 -6.6289 0.6083 -6.6375 0.5074 -6.6224 C 0.3158 -6.5936 0.1198 -6.6216 -0.074 -6.6246 C -0.5726 -6.6323 -1.0725 -6.6396 -1.5704 -6.6689 C -1.7574 -6.6799 -1.1958 -6.6589 -1.0089 -6.6468 C -0.7853 -6.6322 -0.6189 -6.6183 -0.3986 -6.5936 C -0.3288 -6.5858 -0.122 -6.5659 -0.1914 -6.5559 C -0.5392 -6.5062 -0.9228 -6.5515 -1.2736 -6.5515 C -1.3923 -6.5515 -1.7955 -6.5536 -1.3888 -6.5415 C -1.139 -6.5341 -0.8825 -6.5449 -0.6345 -6.5094 C -0.6047 -6.5052 -0.5251 -6.4841 -0.5525 -6.4718 C -0.6065 -6.4475 -0.8106 -6.4344 -0.8283 -6.433 C -1.0755 -6.4129 -1.3433 -6.3791 -1.5915 -6.4031 C -1.7084 -6.4144 -1.3567 -6.4119 -1.2392 -6.4119 C -1.1403 -6.4119 -1.0413 -6.4119 -0.9424 -6.4119 C -0.8523 -6.4119 -0.8546 -6.4115 -0.9346 -6.4131 C -1.2407 -6.4189 -1.5463 -6.4294 -1.8518 -6.4507 C -1.9379 -6.4567 -2.1903 -6.4557 -2.1087 -6.4839 C -1.852 -6.5728 -1.0245 -6.5563 -1.2957 -6.5415 C -1.5894 -6.5256 -1.8872 -6.5285 -2.1807 -6.5449 C -2.306 -6.5518 -2.3901 -6.5582 -2.5085 -6.5814 C -2.5435 -6.5883 -2.6316 -6.5966 -2.6071 -6.6224 C -2.5601 -6.672 -2.4754 -6.6603 -2.4078 -6.67 C -2.2601 -6.6912 -2.1576 -6.6996 -2.0079 -6.7066 C -1.8788 -6.7125 -1.4972 -6.7384 -1.6203 -6.6988 C -1.8078 -6.6385 -2.0534 -6.6612 -2.2461 -6.6645 C -2.2996 -6.6654 -2.455 -6.6684 -2.5252 -6.6822 C -2.5942 -6.6958 -2.5016 -6.71 -2.4964 -6.711 C -2.0908 -6.7898 -0.8453 -6.7364 -1.257 -6.7021 C -1.5196 -6.6802 -1.7928 -6.6849 -2.0555 -6.7032 C -2.1201 -6.7078 -2.2746 -6.7343 -2.1176 -6.7697 C -1.9023 -6.8182 -1.6697 -6.8117 -1.4508 -6.8195 C -1.3493 -6.8232 -1.2477 -6.8254 -1.1462 -6.8295 C -1.1211 -6.8305 -1.1964 -6.8306 -1.2215 -6.8306 C -1.4394 -6.8306 -1.6584 -6.8243 -1.8761 -6.8339 C -1.8782 -6.834 -2.2115 -6.8539 -2.0356 -6.9181 C -1.8558 -6.9838 -1.6351 -6.9636 -1.4486 -6.958 C -1.409 -6.9568 -1.244 -6.9531 -1.1839 -6.9381 C -1.1207 -6.9223 -1.2102 -6.9018 -1.2115 -6.9015 C -1.3764 -6.859 -1.5793 -6.8628 -1.7432 -6.907 C -1.8408 -6.9334 -1.9235 -7.0211 -1.7886 -7.0732 C -1.6212 -7.1378 -1.4039 -7.1097 -1.2293 -7.1097 C -1.1015 -7.1097 -0.7195 -7.1173 -0.846 -7.0998 C -1.0415 -7.0726 -1.2583 -7.0959 -1.4519 -7.1241 C -1.4714 -7.127 -1.8889 -7.1928 -1.649 -7.2482 C -1.5661 -7.2673 -1.3863 -7.2878 -1.309 -7.298 C -1.2294 -7.3085 -1.1404 -7.2954 -1.0709 -7.3357 C -1.0225 -7.3637 -1.1822 -7.3477 -1.2381 -7.3501 C -1.3748 -7.3559 -1.4534 -7.3577 -1.5892 -7.3711 C -1.6896 -7.381 -1.9566 -7.3812 -1.8817 -7.4487 C -1.8005 -7.5217 -1.5674 -7.5154 -1.4685 -7.5262 C -1.3384 -7.5404 -1.2084 -7.5483 -1.0775 -7.5483 C -0.9937 -7.5483 -1.2451 -7.5483 -1.329 -7.5483 C -1.3748 -7.5483 -1.6213 -7.5175 -1.6568 -7.5683 C -1.659 -7.5715 -1.2583 -7.6092 -1.2581 -7.6093 C -1.2265 -7.6116 -0.9048 -7.6557 -0.9103 -7.6569 C -1.0425 -7.6858 -1.1858 -7.6707 -1.3201 -7.6901 C -1.4025 -7.702 -1.5011 -7.7303 -1.3644 -7.7577 C -1.2135 -7.7879 -1.0583 -7.7993 -0.9047 -7.8053 C -0.7331 -7.812 -0.7872 -7.8053 -0.9125 -7.7964 C -1.0087 -7.7896 -1.0704 -7.7849 -1.1673 -7.7787 C -1.2186 -7.7754 -1.3576 -7.8051 -1.3212 -7.7688 C -1.31 -7.7575 -1.2895 -7.768 -1.2736 -7.7676 C -1.2182 -7.7665 -1.1628 -7.7661 -1.1074 -7.7643 C -0.917 -7.7582 -0.7884 -7.7518 -0.5968 -7.7366 C -0.5765 -7.735 -0.3493 -7.7186 -0.2878 -7.699 C -0.266 -7.692 -0.3272 -7.673 -0.3498 -7.6691 C -0.4723 -7.6478 -0.6064 -7.637 -0.7308 -7.6292 C -0.7523 -7.6279 -0.9433 -7.6222 -0.9092 -7.6502 C -0.7698 -7.765 -0.3859 -7.7655 -0.2269 -7.72 C -0.0914 -7.6813 -0.1739 -7.5869 -0.2601 -7.5528 C -0.424 -7.4879 -0.6291 -7.4706 -0.8029 -7.4952 C -0.8931 -7.5079 -1.0196 -7.5578 -0.9324 -7.6657 C -0.8389 -7.7816 -0.6648 -7.782 -0.5304 -7.7843 C -0.4714 -7.7853 -0.2183 -7.8113 -0.1793 -7.7377 C -0.1248 -7.6352 -0.364 -7.5421 -0.4185 -7.5295 C -0.5747 -7.4933 -0.7804 -7.4975 -0.9313 -7.555 C -1.0861 -7.6139 -1.021 -7.7141 -0.9036 -7.771 C -0.7494 -7.8457 -0.562 -7.8604 -0.3986 -7.8075 C -0.2248 -7.7513 -0.1935 -7.6204 -0.3388 -7.4985 C -0.4729 -7.386 -0.6544 -7.3522 -0.815 -7.4243 C -0.8913 -7.4585 -0.9706 -7.5327 -0.9059 -7.617 C -0.8111 -7.7405 -0.5838 -7.7502 -0.4473 -7.761 C -0.1599 -7.7837 0.1366 -7.7598 0.41 -7.6635 C 0.6525 -7.5781 0.6218 -7.4827 0.3967 -7.4132 C 0.1284 -7.3304 -0.1699 -7.3077 -0.4473 -7.3512 C -0.5183 -7.3623 -0.9196 -7.4346 -0.8715 -7.5915 C -0.8359 -7.7079 -0.542 -7.7083 -0.4584 -7.7123 C -0.2629 -7.7215 0.0225 -7.7336 0.1918 -7.6137 C 0.368 -7.4889 -0.06 -7.4126 -0.0984 -7.4066 C -0.3065 -7.3736 -0.5668 -7.3593 -0.7508 -7.4841 C -0.9337 -7.6081 -0.707 -7.6829 -0.599 -7.7001 C -0.4287 -7.7271 -0.2194 -7.7321 -0.0574 -7.6624 C 0.1732 -7.5632 -0.1845 -7.4557 -0.2513 -7.4398 C -0.2925 -7.4299 -0.8537 -7.3245 -0.8305 -7.4863 C -0.8117 -7.6181 -0.3632 -7.567 -0.2967 -7.5616 C -0.2621 -7.5588 0.1908 -7.5316 0.1043 -7.3911 C -0.0067 -7.2108 -0.3736 -7.1498 -0.5625 -7.1352 C -0.6449 -7.1288 -0.8588 -7.1772 -0.7929 -7.2271 C -0.6479 -7.3368 -0.3101 -7.2642 -0.1538 -7.2438 C -0.0267 -7.2271 0.3087 -7.1592 0.1995 -7.092 C 0.0547 -7.0029 -0.1929 -7.0116 -0.3554 -7.0167 C -0.4326 -7.0191 -0.6529 -7.0712 -0.5769 -7.0854 C -0.44 -7.1108 -0.2769 -7.0836 -0.1405 -7.0687 C -0.1075 -7.0652 -0.005 -7.0549 0.0367 -7.0388 C 0.075 -7.0241 0.0549 -7.0112 0.0267 -7.0001 C -0.1083 -6.9466 -0.2707 -6.9567 -0.413 -6.9502 C -0.4604 -6.9481 -0.6201 -6.9303 -0.4916 -6.9303 C -0.3883 -6.9303 -0.6961 -6.8879 -0.7885 -6.8417 C -0.8449 -6.8135 -0.9249 -6.8042 -0.9856 -6.7907 C -1.0999 -6.7653 -1.2301 -6.6991 -1.3356 -6.649 C -1.4699 -6.5852 -1.5591 -6.4924 -1.6845 -6.4219 C -2.0818 -6.1985 -2.2639 -5.7959 -2.4144 -5.3874 C -2.4353 -5.3306 -2.1641 -5.34 -2.1065 -5.3354 C -2.0181 -5.3283 -1.9327 -5.3154 -1.844 -5.3154 C -1.7701 -5.3154 -1.5486 -5.3154 -1.6225 -5.3154 C -1.7481 -5.3154 -1.8685 -5.3553 -1.9946 -5.3553 C -2.108 -5.3553 -2.2213 -5.3553 -2.3347 -5.3553 C -2.4034 -5.3553 -2.5248 -5.3216 -2.4997 -5.4262 C -2.4798 -5.5092 -2.3715 -5.6418 -2.3114 -5.702 C -2.1725 -5.8409 -2.0089 -5.9569 -1.906 -6.1317 C -1.8467 -6.2325 -1.7702 -6.2988 -1.6867 -6.3798 C -1.5537 -6.5089 -1.3154 -6.4949 -1.1429 -6.4906 C -0.687 -6.4792 -0.0718 -6.4913 0.2959 -6.1716 C 0.3841 -6.0949 0.4469 -5.9808 0.5329 -5.898 C 0.6101 -5.8237 0.6946 -5.7433 0.7555 -5.6555 C 0.7981 -5.5941 0.8615 -5.542 0.8951 -5.4749 C 0.9294 -5.4063 0.9325 -5.4151 0.8408 -5.4151 C 0.5595 -5.4151 0.2782 -5.4151 -0.0032 -5.4151 C -0.1031 -5.4151 -0.4021 -5.3952 -0.3022 -5.3952 C -0.1064 -5.3952 0.0888 -5.4151 0.2848 -5.4151 C 0.3864 -5.4151 0.7832 -5.4727 0.8386 -5.3741 C 0.8825 -5.2962 0.9181 -5.233 0.9748 -5.1659 C 1.0122 -5.1218 1.1258 -4.9988 1.0325 -4.9499 C 0.9379 -4.9004 0.8205 -4.9152 0.7544 -4.8137 C 0.7333 -4.7812 0.6891 -4.6728 0.6547 -4.6575 C 0.6192 -4.6417 0.5792 -4.6424 0.5451 -4.6254 C 0.3927 -4.5492 0.233 -4.5539 0.0666 -4.5279 C -0.0436 -4.5107 -0.1411 -4.5245 -0.249 -4.5467 C -0.3083 -4.5589 -0.4519 -4.585 -0.5071 -4.5534 C -0.5612 -4.5225 -0.5093 -4.2111 -0.5093 -4.1402 C -0.5093 -3.9834 -0.3316 -3.9078 -0.1992 -3.88 C -0.1384 -3.8672 -0.0995 -3.8458 -0.0596 -3.798 C -0.0338 -3.767 0.0738 -3.688 0.0478 -3.6928 C 0.0189 -3.6981 -0.2493 -3.7285 -0.1217 -3.6895 C -0.0464 -3.6665 0.0326 -3.6502 0.1109 -3.6429 C 0.1282 -3.6413 0.2107 -3.6385 0.1397 -3.6607 C 0.1081 -3.6705 -0.1878 -3.7022 -0.1051 -3.6319 C -0.0091 -3.5503 0.1771 -3.6731 -0.0153 -3.6961 C -0.061 -3.7016 -0.3074 -3.6663 -0.1837 -3.6695 C -0.1711 -3.6699 -0.0169 -3.6926 -0.0818 -3.7315 C -0.1405 -3.7668 -0.2432 -3.7646 -0.3077 -3.7515 C -0.4354 -3.7256 -0.2887 -3.66 -0.228 -3.6895 C -0.1938 -3.7061 -0.2262 -3.7444 -0.2479 -3.7515 C -0.3026 -3.7694 -0.391 -3.7681 -0.4429 -3.7426 C -0.5569 -3.6865 -0.2644 -3.6779 -0.5005 -3.7116 C -0.556 -3.7195 -0.6294 -3.7223 -0.671 -3.6762 C -0.6861 -3.6595 -0.6795 -3.6158 -0.6932 -3.6053 C -0.7621 -3.5523 -0.9399 -3.6717 -1.0233 -3.6008 C -1.0637 -3.5665 -1.0895 -3.5579 -1.1473 -3.561 C -1.2526 -3.5665 -1.3421 -3.6168 -1.453 -3.5964 C -1.4855 -3.5905 -1.5092 -3.5641 -1.4818 -3.5344 C -1.4781 -3.5304 -1.4293 -3.5074 -1.4331 -3.5056 C -1.4827 -3.4823 -1.5634 -3.5284 -1.5416 -3.4414 C -1.5246 -3.3734 -1.534 -3.4747 -1.5416 -3.4967 C -1.5553 -3.5362 -1.6974 -3.6272 -1.6812 -3.5765 C -1.653 -3.4884 -1.6881 -3.5624 -1.7122 -3.4923 C -1.7366 -3.4212 -1.708 -3.4703 -1.7609 -3.5122 C -1.8123 -3.5529 -1.7911 -3.3323 -1.7764 -3.2951 C -1.7523 -3.2342 -1.6882 -3.2859 -1.7565 -3.211 C -1.8331 -3.127 -1.6764 -3.1431 -1.6922 -3.1046 C -1.7089 -3.0643 -1.837 -3.0061 -1.7343 -3.0027 C -1.6897 -3.0013 -1.7476 -2.9938 -1.721 -2.9673 C -1.687 -2.9333 -1.7132 -2.9522 -1.721 -2.8876 C -1.7286 -2.8259 -1.5909 -2.8515 -1.6347 -2.8233 C -1.6987 -2.782 -1.7236 -2.7263 -1.6347 -2.7236 C -1.5757 -2.7218 -1.6289 -2.6528 -1.597 -2.624 C -1.5661 -2.596 -1.4489 -2.6265 -1.4464 -2.5841 C -1.4427 -2.5214 -1.3902 -2.5582 -1.3445 -2.5353 C -1.2963 -2.5113 -1.3093 -2.4702 -1.2315 -2.4644 C -1.1875 -2.4612 -1.2264 -2.4217 -1.1983 -2.4091 C -1.1057 -2.3674 -0.9794 -2.406 -0.8903 -2.3891 C -0.8105 -2.374 -0.7214 -2.3784 -0.64 -2.3692 C -0.602 -2.3649 -0.4353 -2.4176 -0.5138 -2.4755 C -0.5471 -2.5002 -0.5399 -2.4426 -0.4628 -2.4445 C -0.4274 -2.4454 -0.4478 -2.4552 -0.4407 -2.4755 C -0.434 -2.4943 -0.3498 -2.5065 -0.3343 -2.5132 C -0.2221 -2.5617 -0.0723 -2.5974 0.0533 -2.5597 C 0.085 -2.5502 0.1719 -2.4999 0.204 -2.5287 C 0.2229 -2.5457 0.189 -2.6218 0.1885 -2.6439 C 0.1873 -2.69 0.1935 -2.7386 0.1707 -2.7812 C 0.1611 -2.7993 0.1095 -2.8201 0.1419 -2.8277 C 0.1481 -2.8292 0.2956 -2.8534 0.235 -2.8787 C 0.1719 -2.905 0.182 -2.9581 0.1309 -3.0116 C 0.0806 -3.0643 0.0261 -3.0914 0.029 -3.1667 C 0.0321 -3.2491 0.0888 -3.3116 0.0888 -3.4015 C 0.0888 -3.4324 0.1193 -3.4982 0.0998 -3.5211 C 0.0784 -3.5463 0.0425 -3.5785 0.0201 -3.6008 C -0.0217 -3.6426 0.0701 -3.481 0.1087 -3.4547 C 0.214 -3.3829 0.1632 -3.4201 0.0976 -3.3372 C 0.0258 -3.2464 0.1181 -3.2719 0.1441 -3.222 C 0.1713 -3.1702 0.1515 -3.0681 0.1641 -3.0072 C 0.1737 -2.9606 0.1087 -2.9088 0.1087 -2.8543 C 0.1087 -2.782 0.1299 -2.712 0.1286 -2.6372 C 0.1262 -2.4889 0.1535 -2.5655 0.1486 -2.6638 C 0.1417 -2.8019 0.0714 -2.9868 -0.0464 -3.0626 C -0.2763 -3.2104 -0.5606 -3.2207 -0.8261 -3.2575 C -1.1633 -3.3042 -1.4923 -3.2005 -1.7875 -3.047 C -1.861 -3.0088 -1.5721 -2.9571 -1.5305 -2.9474 C -1.384 -2.913 -1.2488 -2.9031 -1.0986 -2.9031 C -0.969 -2.9031 -0.8168 -2.8812 -0.6888 -2.9075 C -0.613 -2.923 -0.5374 -2.9488 -0.4606 -2.9584 C -0.2967 -2.9789 -0.1402 -2.9835 0.0157 -3.047 C 0.0878 -3.0765 -0.0169 -3.1216 -0.0464 -3.1224 C -0.1444 -3.125 -0.2429 -3.1224 -0.341 -3.1224 C -0.7264 -3.1224 -1.1119 -3.1224 -1.4973 -3.1224 C -1.553 -3.1224 -1.6815 -3.1474 -1.721 -3.098 C -1.7562 -3.0541 -1.78 -2.9114 -1.7521 -2.8632 C -1.7136 -2.7968 -1.5622 -2.7796 -1.4951 -2.8078 C -1.4824 -2.8132 -1.5946 -2.7944 -1.5394 -2.7259 C -1.4607 -2.6281 -1.3841 -2.7825 -1.5239 -2.6993 C -1.6404 -2.6299 -1.3825 -2.6266 -1.3644 -2.6284 C -1.2733 -2.6373 -1.3159 -2.6441 -1.3068 -2.5641 C -1.3026 -2.5274 -1.2071 -2.4781 -1.2071 -2.5243 C -1.2071 -2.5339 -1.2482 -2.4861 -1.2027 -2.4844 C -1.1632 -2.4829 -1.1308 -2.5062 -1.1318 -2.4999 C -1.1457 -2.4165 -0.9361 -2.3953 -0.928 -2.48 C -0.9187 -2.5778 -1.0393 -2.5307 -0.9989 -2.4224 C -0.9603 -2.3187 -0.7664 -2.4662 -0.8372 -2.5464 C -0.8816 -2.5968 -0.9928 -2.4291 -0.8483 -2.4334 C -0.7588 -2.4361 -0.7176 -2.5684 -0.7331 -2.5509 C -0.82 -2.4521 -0.6271 -2.4777 -0.5824 -2.5088 C -0.4864 -2.5755 -0.5754 -2.573 -0.5115 -2.5287 C -0.4249 -2.4685 -0.3798 -2.5234 -0.3498 -2.5907 C -0.3413 -2.61 -0.3181 -2.5618 -0.2989 -2.5531 C -0.2687 -2.5393 -0.2394 -2.5535 -0.2191 -2.5774 C -0.1543 -2.6541 -0.1476 -2.5781 -0.0397 -2.5686 C 0.0362 -2.5619 0.0329 -2.6546 0.0578 -2.6328 C 0.0921 -2.6028 0.0822 -2.6916 0.0334 -2.7148 C -0.0335 -2.7465 0.0188 -2.7834 0.0334 -2.8477 C 0.0354 -2.8565 0.0675 -2.954 0.0334 -2.954 C 0.0278 -2.954 0.0636 -2.9286 0.0688 -2.9274 C 0.0949 -2.9219 0.123 -2.9361 0.1154 -2.9673 C 0.0965 -3.0448 -0.01 -3.0079 0.0245 -3.0603 C 0.1046 -3.1822 -0.0088 -3.3315 0.0489 -3.4502 C 0.0974 -3.55 0.0253 -3.5057 -0.0065 -3.5455 C -0.0306 -3.5756 0.0409 -3.6224 -0.0198 -3.6518 C -0.0758 -3.679 -0.0843 -3.6082 -0.135 -3.6341 C -0.1674 -3.6506 -0.3349 -3.7145 -0.2856 -3.6097 C -0.2831 -3.6043 -0.262 -3.6208 -0.2989 -3.6053 C -0.3263 -3.5937 -0.3929 -3.5704 -0.4185 -3.5809 C -0.5396 -3.6308 -0.1934 -3.5377 -0.1726 -3.5211 C -0.1559 -3.5077 -0.2154 -3.5211 -0.2369 -3.5211 C -0.2524 -3.5211 -0.2058 -3.5211 -0.1903 -3.5211 C -0.1573 -3.5211 -0.1183 -3.5157 -0.0862 -3.5255 C 0.086 -3.5785 0.0337 -3.6236 -0.1062 -3.7116 C -0.1585 -3.7445 -0.5077 -3.9963 -0.5802 -3.911 C -0.6195 -3.8648 -0.4956 -3.7636 -0.5337 -3.7205 C -0.6043 -3.6404 -0.9017 -3.8497 -0.9856 -3.8578 C -1.0616 -3.8652 -0.9549 -3.7447 -0.9457 -3.7382 C -0.935 -3.7305 -0.7486 -3.6601 -0.7486 -3.685 C -0.7486 -3.7106 -1.0902 -3.731 -1.103 -3.7315 C -1.1082 -3.7318 -1.5011 -3.7168 -1.3356 -3.6274 C -1.2957 -3.6059 -1.0873 -3.6124 -1.1229 -3.6319 C -1.16 -3.6521 -1.3138 -3.6785 -1.3334 -3.6186 C -1.3847 -3.4618 -0.9945 -3.5623 -1.1983 -3.592 C -1.2124 -3.594 -1.396 -3.6027 -1.3799 -3.5543 C -1.3748 -3.5391 -1.3034 -3.5327 -1.3002 -3.5499 C -1.2876 -3.6168 -1.5587 -3.5373 -1.5372 -3.5543 C -1.4235 -3.6445 -1.5169 -3.5975 -1.597 -3.5831 C -1.6143 -3.58 -1.5587 -3.6173 -1.5549 -3.6252 C -1.5343 -3.6682 -1.6854 -3.6363 -1.6856 -3.6363 C -1.7328 -3.6263 -1.8668 -3.5771 -1.8185 -3.5787 C -1.7687 -3.5804 -1.7058 -3.6225 -1.6679 -3.6518 C -1.6283 -3.6823 -1.6979 -3.6749 -1.6524 -3.7293 C -1.5623 -3.837 -1.4862 -3.8689 -1.4862 -4.0195 C -1.4862 -4.0877 -1.5127 -4.1615 -1.4973 -4.2278 C -1.4823 -4.2925 -1.4415 -4.3544 -1.4464 -4.4227 C -1.4572 -4.575 -1.5929 -4.5641 -1.6856 -4.5069 C -1.8452 -4.4082 -2.2082 -4.3919 -2.359 -4.5135 C -2.3872 -4.5362 -2.4858 -4.5688 -2.5097 -4.5268 C -2.5536 -4.4494 -2.4641 -4.2464 -2.4631 -4.1547 C -2.4607 -3.9141 -2.4539 -3.6803 -2.4277 -3.4414 C -2.3997 -3.1861 -2.3663 -2.8824 -2.4078 -2.6284 C -2.436 -2.4554 -2.4528 -2.2409 -2.5806 -2.1056 C -2.7067 -1.972 -2.8288 -1.8409 -2.9084 -1.6714 C -2.9348 -1.6151 -2.9536 -1.5541 -2.9771 -1.4964 C -3.0034 -1.4316 -2.9303 -1.4217 -2.8796 -1.4012 C -2.6931 -1.3255 -2.5017 -1.2628 -2.3147 -1.1885 C -2.3123 -1.1875 -2.0927 -1.08 -2.0821 -1.0932 C -2.0727 -1.1051 -2.3445 -1.1881 -2.3745 -1.1885 C -2.4663 -1.1898 -2.5877 -1.1357 -2.6714 -1.1907 C -2.8404 -1.3018 -2.9427 -1.573 -2.904 -1.7667 C -2.8853 -1.8602 -2.8197 -1.972 -2.8685 -2.0679 C -2.923 -2.1749 -3.0943 -2.2122 -3.1942 -2.2496 C -3.3081 -2.2922 -3.4354 -2.3271 -3.5375 -2.3958 C -3.5586 -2.41 -3.5609 -2.427 -3.5885 -2.4201 C -3.7421 -2.3817 -3.4255 -2.1918 -3.3957 -2.1743 C -3.1592 -2.035 -2.808 -1.9154 -2.6559 -1.6714 C -2.5961 -1.5756 -2.627 -1.4265 -2.6891 -1.3391 C -2.7095 -1.3103 -2.754 -1.4662 -2.7555 -1.4698 C -2.8069 -1.5918 -2.8688 -1.7088 -2.9372 -1.822 C -3.0599 -2.0252 -3.2194 -2.2047 -3.4533 -2.2762 C -3.4825 -2.2851 -3.5029 -2.2937 -3.4666 -2.2762 C -3.3373 -2.2138 -3.2126 -2.125 -3.1056 -2.0303 C -2.9211 -1.8669 -2.8573 -1.6695 -2.7378 -1.4676 C -2.6963 -1.3973 -2.6824 -1.3639 -2.6824 -1.2837 C -2.6824 -1.2448 -2.6668 -1.1586 -2.7112 -1.1376 C -2.9012 -1.0476 -3.2288 -1.262 -3.378 -1.3502 C -3.6123 -1.4886 -3.8718 -1.5701 -4.1312 -1.6493 C -4.2081 -1.6727 -4.0022 -1.541 -3.923 -1.5274 C -3.7722 -1.5016 -3.6263 -1.4307 -3.4844 -1.3768 C -3.2408 -1.2843 -2.9799 -1.25 -2.7245 -1.2084 C -2.5721 -1.1836 -2.7486 -1.1298 -2.8021 -1.1243 C -3.018 -1.1018 -3.3311 -1.0653 -3.4622 -0.8673 C -3.4718 -0.8528 -3.4736 -0.7969 -3.4976 -0.7942 C -3.5939 -0.7833 -3.7186 -0.8573 -3.8033 -0.8894 C -3.9677 -0.9518 -4.1331 -0.9993 -4.304 -1.0401 C -4.3299 -1.0463 -4.4521 -1.0854 -4.3461 -1.0556 C -4.1868 -1.0108 -4.0248 -0.9169 -3.8609 -0.8939 C -3.594 -0.8563 -3.342 -0.8673 -3.0923 -0.7388 C -3.0921 -0.7387 -2.9855 -0.6513 -3.0236 -0.6502 C -3.2088 -0.6448 -3.4466 -0.6965 -3.6239 -0.6413 C -3.7551 -0.6005 -3.6605 -0.255 -3.5885 -0.185 C -3.5739 -0.1708 -3.5309 -0.0942 -3.5641 -0.0765 C -3.687 -0.0109 -4.5885 -0.4999 -4.5964 -0.3822 C -4.5989 -0.3438 -4.4622 -0.412 -4.3992 -0.4021 C -4.1946 -0.3699 -3.9581 -0.2732 -3.79 -0.154 C -3.6867 -0.0807 -3.5316 0.0655 -3.3957 0.0786 C -3.3343 0.0846 -3.3432 0.0599 -3.3559 0.1362 C -3.3757 0.2549 -3.3987 0.379 -3.4002 0.4995 C -3.4013 0.5928 -3.3706 0.7424 -3.4135 0.8296 C -3.5393 1.0854 -3.9861 0.8827 -4.1533 0.8229 C -4.3142 0.7653 -4.5297 0.6052 -4.7116 0.6546 C -4.8524 0.6927 -4.4629 0.6902 -4.4302 0.6944 C -4.278 0.7142 -4.1484 0.8205 -4.0005 0.865 C -3.9286 0.8866 -3.724 0.9646 -3.7812 0.9159 C -3.8389 0.8669 -3.9261 0.8432 -3.9894 0.8008 C -4.1511 0.6925 -4.2038 0.5014 -4.3173 0.3511 C -4.3819 0.2654 -4.6288 0.077 -4.5875 0.1761 C -4.5534 0.258 -4.3865 0.3367 -4.3217 0.3821 C -4.1999 0.4674 -4.0671 0.5517 -3.9761 0.6723 C -3.8875 0.7896 -3.936 0.9356 -3.8654 1.0533 C -3.8502 1.0786 -3.7502 1.1811 -3.748 1.1795 C -3.7236 1.1627 -3.6292 1.0937 -3.6483 1.133 C -3.7424 1.3268 -3.6607 1.5938 -3.7834 1.7732 C -3.8434 1.8609 -3.9823 1.6852 -3.9983 1.6448 C -4.098 1.3932 -4.2393 1.1052 -4.5277 1.0378 C -4.729 0.9908 -4.6916 1.0617 -4.5632 1.164 C -4.3675 1.3199 -4.2138 1.4759 -4.0869 1.6935 C -4.0218 1.8051 -3.9964 1.9545 -3.903 2.0479 C -3.8765 2.0745 -3.7737 2.0599 -3.8875 2.0612 C -4.17 2.0645 -4.4111 2.206 -4.6806 2.2739 C -4.7774 2.2982 -4.7404 2.2694 -4.6806 2.2694 C -4.5332 2.2694 -4.3826 2.3046 -4.2375 2.2606 C -4.1733 2.2411 -4.0574 2.174 -4.109 2.09 C -4.1256 2.0631 -4.1337 2.0661 -4.1046 2.0944 C -4.037 2.1605 -3.9097 2.2287 -3.8831 2.3248 C -3.8586 2.4134 -3.793 2.5154 -3.8388 2.6039 C -3.9321 2.7845 -4.0708 2.8989 -4.2641 2.9429 C -4.3097 2.9532 -4.2713 2.9574 -4.2464 2.9539 C -4.1409 2.9394 -4.0326 2.9074 -3.9473 2.841 C -3.8357 2.7541 -3.7701 2.6213 -3.7369 2.4865 C -3.7328 2.4698 -3.7195 2.4021 -3.728 2.3758 C -3.7311 2.3663 -3.7501 2.3802 -3.7546 2.3891 C -3.8222 2.5215 -3.7728 2.7056 -3.8255 2.8498 C -3.888 3.0211 -3.947 3.1426 -3.9584 3.3217 C -3.9667 3.4528 -4.1071 3.4343 -4.1445 3.5277 C -4.1629 3.5736 -4.0504 3.4964 -4.0049 3.4767 C -3.8962 3.4296 -3.7929 3.3765 -3.7325 3.2685 C -3.5842 3.0034 -3.757 2.72 -3.7945 2.4533 C -3.8054 2.3754 -3.3053 2.4757 -3.2473 2.533 C -3.0996 2.679 -3.25 3.0364 -3.3293 3.1688 C -3.4077 3.2997 -3.5076 3.3915 -3.4755 3.552 C -3.4438 3.7104 -3.3724 3.4115 -3.3692 3.3903 C -3.3413 3.21 -3.3689 3.0374 -3.3913 2.8587 C -3.4057 2.744 -3.4074 2.6274 -3.3249 2.5353 C -3.254 2.4562 -3.0587 2.4451 -3.0014 2.5441 C -2.8949 2.7284 -2.9273 2.9596 -2.9217 3.1622 C -2.9152 3.3951 -2.7201 3.5328 -2.5584 3.6739 C -2.4832 3.7394 -2.5193 3.676 -2.5451 3.6207 C -2.6326 3.433 -2.7519 3.2629 -2.8486 3.0802 C -2.9233 2.9389 -2.9587 2.7769 -2.9616 2.6172 C -2.9623 2.5769 -2.9416 2.4749 -2.966 2.4444 C -2.9725 2.4363 -2.9818 2.5789 -2.9815 2.5951 C -2.9787 2.7232 -2.876 2.77 -2.7666 2.8077 C -2.696 2.8321 -2.4792 2.8786 -2.554 2.8786 C -2.6573 2.8786 -2.772 2.8366 -2.8641 2.7944 C -3.0389 2.7144 -3.0762 2.592 -3.0125 2.4156 C -2.9302 2.1875 -2.7275 2.0214 -2.4919 1.9859 C -2.4215 1.9753 -2.2944 2.0376 -2.2372 2.0745 C -2.1272 2.1454 -2.2628 2.4046 -2.2859 2.4821 C -2.356 2.7169 -2.1244 2.8615 -2.0578 3.0647 C -2.0129 3.2016 -2.0573 3.5266 -2.1929 3.6118 C -2.2237 3.6312 -2.4628 3.7074 -2.3369 3.6384 C -2.0167 3.4631 -2.2129 3.1237 -2.266 2.8653 C -2.2897 2.7501 -2.2789 2.6338 -2.2881 2.5175 C -2.2948 2.4336 -2.3281 2.3703 -2.3502 2.2916 C -2.3758 2.2001 -2.3609 2.1118 -2.3479 2.0213 C -2.3321 1.9105 -3.1359 2.07 -2.5473 1.822 C -2.3721 1.7482 -2.0409 1.5915 -2.0578 1.3523 C -2.0717 1.1537 -2.2348 1.614 -2.2372 1.6182 C -2.3113 1.7467 -2.42 1.9338 -2.565 1.997 C -2.6177 2.0199 -2.5709 1.976 -2.5606 1.9682 C -2.2997 1.7694 -2.184 1.4698 -2.184 1.1452 C -2.184 0.942 -2.1922 0.85 -2.3756 0.7443 C -2.4167 0.7206 -2.6823 0.5399 -2.7024 0.6501 C -2.755 0.9396 -2.8335 1.247 -2.9328 1.5229 C -2.9437 1.5533 -3.1788 1.9208 -3.1842 1.7145 C -3.1875 1.5873 -3.0621 1.4496 -3.0025 1.349 C -2.914 1.1995 -2.8286 1.0425 -2.8286 0.8639 C -2.8286 0.7883 -2.8468 0.7104 -2.925 0.6822 C -2.9572 0.6707 -2.8651 0.7169 -2.832 0.7254 C -2.6678 0.7676 -2.4286 0.7568 -2.3491 0.9281 C -2.2777 1.0819 -2.1744 1.1886 -2.009 1.2316 C -1.857 1.2711 -1.7766 1.2481 -1.6335 1.1851 C -1.4976 1.1252 -1.3508 1.1035 -1.2138 1.1762 C -1.0502 1.2631 -1.2568 1.7698 -1.2647 1.9039 C -1.2805 2.1715 -1.2492 2.4386 -1.2492 2.7058 C -1.2492 2.9139 -1.3074 3.1164 -1.3356 3.3217 C -1.3417 3.3659 -1.3547 3.5057 -1.3456 3.3826 C -1.328 3.1453 -1.2427 2.9347 -1.2603 2.6881 C -1.2802 2.4088 -1.3068 2.1408 -1.3068 1.8585 C -1.3068 1.7486 -1.334 1.7023 -1.2149 1.7023 C -1.1539 1.7023 -1.0543 1.6496 -1.0022 1.6669 C -0.8942 1.7029 -0.8351 1.8492 -0.7873 1.9383 C -0.7045 2.0926 -0.6121 2.2422 -0.5304 2.3968 C -0.4566 2.5363 -0.4018 2.6884 -0.3664 2.8421 C -0.3588 2.8754 -0.3593 2.975 -0.3476 2.9429 C -0.3053 2.8264 -0.3202 2.7524 -0.3498 2.6338 C -0.3991 2.4366 -0.359 2.2368 -0.3952 2.0413 C -0.4194 1.9111 -0.5446 1.8454 -0.65 1.7854 C -0.7332 1.7381 -0.9747 1.5892 -0.907 1.6569 C -0.8748 1.6891 -0.7795 1.709 -0.7408 1.73 C -0.5867 1.8137 -0.4491 1.9264 -0.3277 2.0523 C -0.1761 2.2095 -0.0309 2.389 0.0655 2.5862 C 0.1002 2.6572 0.1063 2.7457 0.1486 2.8122 C 0.1701 2.8459 0.1897 2.8339 0.1818 2.7955 C 0.1503 2.642 0.0478 2.499 -0.0331 2.3691 C -0.2245 2.0615 -0.3562 1.748 -0.5071 1.421 C -0.5789 1.2654 -0.6375 1.1416 -0.5957 0.9669 C -0.5621 0.8263 -0.392 0.7499 -0.2712 0.6955 C -0.111 0.6235 0.1202 0.6744 0.266 0.5637 C 0.5187 0.3719 0.3797 0.0294 0.3435 -0.2204 C 0.3392 -0.25 0.3539 -0.5036 0.4476 -0.4165 C 0.6974 -0.1846 0.1014 0.4199 0.0334 0.6579 C 0.0083 0.7456 0.1702 0.825 0.2217 0.8716 C 0.2813 0.9257 0.2682 0.9906 0.2682 1.0644 C 0.2682 1.5074 0.5058 1.9617 0.3989 2.4046 C 0.3663 2.5396 0.4089 2.4001 0.4288 2.3381 C 0.546 1.9729 0.4711 1.685 0.3612 1.3302 C 0.3201 1.1974 0.2803 1.0651 0.235 0.9337 C 0.1955 0.8192 0.1171 0.6563 0.1574 0.5327 C 0.2073 0.3801 0.4653 0.5342 0.534 0.5759 C 0.7652 0.7164 1.0075 0.8376 0.9438 1.1497 C 0.8973 1.3778 0.8929 1.6084 0.8874 1.8419 C 0.8823 2.056 0.8354 2.2453 0.7666 2.4478 C 0.7416 2.5216 0.6223 2.7356 0.6935 2.5286 C 0.8218 2.1553 1.0598 1.8067 0.7799 1.4321 C 0.7387 1.3769 0.7129 1.3569 0.6681 1.3136 C 0.6596 1.3054 0.6443 1.2789 0.6415 1.2903 C 0.6303 1.3348 0.7778 1.3962 0.7932 1.4022 C 0.9266 1.4543 1.096 1.5 1.2407 1.4952 C 1.3196 1.4926 1.3814 1.4248 1.44 1.38 C 1.5332 1.3089 1.631 1.2439 1.7225 1.1707 C 1.758 1.1423 1.794 1.0913 1.7568 1.1674 C 1.5918 1.5049 1.4141 1.8387 1.1897 2.141 C 1.0597 2.3162 0.881 2.4249 0.7412 2.5862 C 0.6777 2.6594 0.6457 2.7515 0.5662 2.8188 C 0.5217 2.8564 0.6361 2.7251 0.678 2.6848 C 0.8452 2.5238 1.0375 2.3919 1.2052 2.2318 C 1.3647 2.0796 1.4514 1.8833 1.6394 1.7566 C 1.752 1.6808 1.8241 1.5841 1.9196 1.4886 C 2.0239 1.3843 2.0826 1.516 2.1312 1.6182 C 2.2931 1.9587 2.172 2.3281 1.9296 2.594 C 1.7336 2.8089 1.4468 2.9565 1.2883 3.2042 C 1.2124 3.3229 1.5711 3.0861 1.5885 3.0736 C 1.8411 2.8912 1.8634 2.5716 1.975 2.2993 C 2.0722 2.0623 2.2281 1.8601 2.3095 1.6159 C 2.3417 1.5195 2.3666 1.276 2.5288 1.349 C 2.6689 1.4121 2.7885 1.5494 2.9043 1.647 C 2.9787 1.7097 3.1085 1.7846 3.1325 1.8884 C 3.2027 2.1927 2.9526 2.5454 2.7813 2.769 C 2.6826 2.8979 2.5586 3.0014 2.4402 3.1112 C 2.4161 3.1336 2.2924 3.2256 2.3007 3.2751 C 2.3125 3.3461 2.5243 3.3292 2.5499 3.2618 C 2.5815 3.1786 2.5911 3.0863 2.6196 3.0016 C 2.6855 2.806 2.7744 2.6203 2.8533 2.43 C 2.9388 2.2241 2.9946 1.9968 2.9785 1.7721 C 2.9753 1.7269 2.8653 1.7016 2.9896 1.7057 C 3.0828 1.7087 3.459 1.6958 3.4802 1.8231 C 3.5364 2.1598 3.4808 2.5462 3.6032 2.8664 C 3.6758 3.0565 3.8055 3.2045 3.7859 3.4191 C 3.7742 3.5488 3.7771 3.1353 3.7771 3.1156 C 3.7771 2.8977 3.7816 2.6764 3.7284 2.4633 C 3.6801 2.2701 3.5868 2.0923 3.5334 1.9006 C 3.5079 1.8089 3.4647 1.6924 3.4647 1.596 C 3.4647 1.5443 3.559 1.6436 3.5943 1.6813 C 3.7987 1.8995 3.9086 2.1002 4.0075 2.3769 C 4.0687 2.5482 4.1805 2.7065 4.3264 2.8166 C 4.3853 2.861 4.7234 3.0634 4.7883 2.9661 C 4.7986 2.9508 4.7697 2.9331 4.7551 2.9218 C 4.6912 2.8725 4.5647 2.7914 4.5081 2.7535 C 4.3588 2.6532 4.1721 2.5553 4.1016 2.3791 C 4.0548 2.262 4.0301 2.1408 3.9753 2.0258 C 3.9359 1.943 3.8832 1.8667 3.848 1.7821 C 3.7576 1.5653 3.5075 0.9292 3.9897 0.9614 C 4.3005 0.9821 4.4118 1.2019 4.5325 1.4553 C 4.6122 1.6229 4.6753 1.7042 4.8559 1.7422 C 4.8882 1.749 4.9388 1.8284 4.9633 1.8508 C 4.9833 1.869 5.0485 1.9066 4.9744 1.8519 C 4.81 1.7304 4.6198 1.6383 4.4428 1.5373 C 4.3414 1.4795 3.9622 1.3398 3.9355 1.1928 C 3.9105 1.0555 4.1472 1.1244 4.2334 1.1408 C 4.3067 1.1547 4.8211 1.356 4.8636 1.2803 C 4.9458 1.1343 4.7563 0.9703 4.6853 0.8617 C 4.6067 0.7414 4.5393 0.5096 4.3951 0.4463 C 4.2342 0.3757 4.036 0.3666 3.8624 0.3666 C 3.7642 0.3666 3.7746 0.2354 3.7826 0.1561 C 3.8024 -0.0387 4.0326 -0.1692 4.1625 -0.2902 C 4.2721 -0.3923 4.3438 -0.4713 4.3541 -0.6214 C 4.3569 -0.6616 4.3467 -0.7183 4.3652 -0.7554 C 4.4007 -0.8265 4.3762 -0.5966 4.373 -0.5173 C 4.3664 -0.3576 4.3347 0.1223 4.0562 -0.01 C 3.9243 -0.0727 3.7827 -0.111 3.6364 -0.1163 C 3.543 -0.1198 3.5157 -0.1141 3.4481 -0.1817 C 3.3115 -0.3183 3.2118 -0.4859 3.0848 -0.6314 C 3.0758 -0.6418 2.9686 -0.7253 3.0228 -0.7288 C 3.1552 -0.7374 3.2813 -0.7449 3.4094 -0.7864 C 3.6916 -0.878 3.7088 -1.1465 3.7162 -1.4078 C 3.7178 -1.4656 3.7586 -1.5933 3.6951 -1.6171 C 3.5454 -1.6733 3.7524 -1.5967 3.7948 -1.5618 C 4.0075 -1.386 3.8421 -1.2312 3.6287 -1.2029 C 3.5055 -1.1866 3.3293 -1.224 3.2155 -1.1697 C 3.1154 -1.1219 3.2328 -0.9268 3.2609 -0.875 C 3.3516 -0.7082 3.4497 -0.6207 3.1956 -0.6136 C 3.1065 -0.6112 2.9633 -0.5862 2.9287 -0.6801 C 2.8869 -0.7934 2.7983 -0.9216 2.7127 -1.0046 C 2.5991 -1.1148 2.6314 -1.279 2.7071 -1.4089 C 2.8422 -1.6404 3.0914 -1.7874 3.2676 -1.986 C 3.3259 -2.0517 3.4098 -2.1259 3.4481 -2.2064 C 3.4634 -2.2384 3.3115 -2.2262 3.3086 -2.2263 C 3.2261 -2.2285 2.9797 -2.2521 3.0616 -2.2418 C 3.1631 -2.2291 3.2637 -2.2039 3.3617 -2.1754 C 3.4128 -2.1605 3.5565 -2.1327 3.4825 -2.0524 C 3.3959 -1.9586 3.2538 -1.9141 3.1435 -1.8564 C 2.8929 -1.7252 2.8184 -1.5404 2.7116 -1.2926 C 2.7004 -1.2667 2.6208 -1.0508 2.6208 -1.1088 C 2.6208 -1.1729 2.7177 -1.0261 2.7592 -0.9991 C 2.923 -0.8925 3.1934 -0.7854 3.3939 -0.8374 C 3.584 -0.8867 3.6424 -1.225 3.6962 -1.3812 C 3.7116 -1.4259 3.773 -1.5516 3.7427 -1.5152 C 3.6969 -1.4602 3.6962 -1.333 3.6896 -1.2682 C 3.6742 -1.118 3.7494 -0.9354 3.5622 -0.9027 C 3.4432 -0.8819 3.3268 -0.9107 3.2344 -0.8285 C 3.0829 -0.6939 2.8098 -0.7881 2.6518 -0.8562 C 2.5214 -0.9124 2.5236 -0.9834 2.4923 -1.1132 C 2.4722 -1.1962 2.3924 -1.1962 2.3128 -1.2073 C 2.1899 -1.2245 2.1105 -1.2677 1.9994 -1.3192 C 1.7014 -1.4572 1.8386 -1.6899 2.0271 -1.8763 C 2.1838 -2.0313 2.3815 -2.1471 2.3815 -2.3836 C 2.3815 -2.4684 2.3585 -2.5279 2.4535 -2.5553 C 2.5311 -2.5776 2.645 -2.5818 2.7116 -2.6306 C 2.7153 -2.6333 2.6219 -2.6379 2.6019 -2.6273 C 2.5162 -2.5817 2.4746 -2.4524 2.4424 -2.3703 C 2.3871 -2.229 2.3387 -2.0854 2.2807 -1.945 C 2.2053 -1.7624 2.0925 -1.3852 1.8465 -1.3735 C 1.5628 -1.3599 2.0283 -1.8155 2.0603 -1.8497 C 2.2113 -2.0111 2.4769 -2.3378 2.4169 -2.5852 C 2.3908 -2.6931 2.0409 -2.2287 2.0093 -2.1709 C 1.9272 -2.0203 1.9344 -1.887 1.9418 -1.7212 C 1.9456 -1.636 1.9342 -1.5718 1.8942 -1.4975 C 1.8441 -1.4046 1.9975 -1.374 2.057 -1.358 C 2.1311 -1.3379 2.5942 -1.2486 2.5942 -1.1652 C 2.5942 -1.0666 2.4167 -0.9954 2.3638 -0.9227 C 2.274 -0.7993 2.2423 -0.753 2.2619 -0.5727 C 2.302 -0.2038 2.516 0.1174 2.5554 0.4818 C 2.5667 0.5863 2.5392 0.6517 2.4324 0.6789 C 2.3771 0.693 2.3038 0.6953 2.2973 0.6224 C 2.2822 0.4509 2.5211 0.3563 2.5067 0.1783 C 2.4802 -0.1483 2.122 -0.3076 1.9905 -0.576 C 1.9083 -0.7439 2.0767 -0.7779 2.1821 -0.6668 C 2.3745 -0.4642 2.4468 -0.0838 2.4369 0.1871 C 2.4313 0.3399 2.4075 0.4915 2.3848 0.6424 C 2.3788 0.6823 2.398 0.5626 2.4048 0.5228 C 2.4402 0.3145 2.4866 0.0913 2.4679 -0.1208 C 2.4498 -0.3257 2.3263 -0.4813 2.1799 -0.6159 C 2.0953 -0.6937 1.8936 -0.7905 2.1223 -0.8075 C 2.156 -0.81 2.2355 -0.8272 2.2586 -0.792 C 2.355 -0.645 2.4152 -0.4695 2.4978 -0.3146 C 2.6179 -0.0894 2.747 0.0728 2.9475 0.227 C 3.1172 0.3576 3.2221 0.4801 3.1657 0.6977 C 3.1494 0.7604 3.0859 0.8626 3.0859 0.927 C 3.0859 0.9678 3.1577 0.8805 3.1757 0.8439 C 3.3138 0.5624 2.9881 0.1795 2.8478 -0.041 C 2.8078 -0.1038 2.6802 -0.3624 2.5931 -0.3811 C 2.5691 -0.3862 2.5996 -0.3319 2.6086 -0.3091 C 2.6374 -0.2357 2.6606 -0.1917 2.7005 -0.1197 C 2.8175 0.0913 2.9497 0.2934 3.0693 0.5028 C 3.1284 0.6062 3.2044 0.7206 3.2166 0.8428 C 3.2212 0.889 3.1628 1.0683 3.1601 1.0112 C 3.1455 0.6957 3.2339 0.298 3.045 0.0199 C 2.9445 -0.128 2.7166 -0.2198 2.6584 -0.38 C 2.6259 -0.4694 2.2454 -0.4885 2.16 -0.4885 C 2.0213 -0.4885 1.9167 -0.1218 1.8632 -0.0244 C 1.7001 0.2724 1.4455 0.472 1.1709 0.6634 C 0.9975 0.7842 1.4591 0.3543 1.604 0.2004 C 1.7563 0.0387 1.7834 -0.2076 1.7834 -0.422 C 1.7834 -0.4509 1.7959 -0.5317 1.779 -0.5084 C 1.7205 -0.4281 1.6797 -0.312 1.6438 -0.2216 C 1.5714 -0.039 1.5294 0.1485 1.4843 0.3389 C 1.4578 0.4511 1.4315 0.5846 1.3293 0.6546 C 1.2222 0.7278 1.3331 0.575 1.3404 0.5671 C 1.5023 0.3892 1.7007 0.2521 1.8842 0.0985 C 2.0301 -0.0235 2.2352 -0.2834 2.0415 -0.4652 C 1.9226 -0.5768 1.53 -0.7557 1.4345 -0.5538 C 1.3087 -0.2879 1.2995 0.0592 1.1044 0.2902 C 1.0034 0.4098 0.9413 0.4233 0.9837 0.2536 C 1.0585 -0.0453 1.2024 -0.3016 1.234 -0.6136 C 1.2378 -0.6512 1.2487 -0.7632 1.2396 -0.7266 C 1.2024 -0.5779 1.1627 -0.4298 1.1288 -0.2803 C 1.0891 -0.105 0.9491 0.4156 1.0014 0.2436 C 1.0844 -0.0289 1.3864 -0.465 1.0945 -0.6945 C 0.9286 -0.8249 0.7118 -0.8707 0.5086 -0.9038 C 0.3965 -0.9221 0.3679 -0.9261 0.3679 -0.8119 C 0.3679 -0.5233 0.2998 -0.4042 0.0278 -0.3002 C -0.0823 -0.2581 -0.2013 -0.2229 -0.2679 -0.1185 C -0.328 -0.0243 -0.3684 0.0851 -0.4097 0.1883 C -0.4522 0.2947 -0.4923 0.4035 -0.5525 0.5017 C -0.5803 0.547 -0.4955 0.4118 -0.4628 0.3699 C -0.3077 0.1708 -0.0891 -0.1042 -0.1139 -0.3777 C -0.1185 -0.4275 -0.1731 -0.2965 -0.1948 -0.2515 C -0.2552 -0.1255 -0.3152 0.0006 -0.3664 0.1307 C -0.4131 0.2488 -0.4598 0.3777 -0.537 0.4807 C -0.5694 0.5238 -0.5177 0.3744 -0.5038 0.3223 C -0.4531 0.1328 -0.3599 -0.1236 -0.4296 -0.319 C -0.4861 -0.4776 -0.7102 -0.5067 -0.8516 -0.5239 C -1.0178 -0.5442 -1.12 -0.5188 -1.1805 -0.3523 C -1.2794 -0.0805 -1.2115 0.169 -1.2115 0.443 C -1.2115 0.5254 -1.3114 0.7165 -1.2913 0.5238 C -1.2595 0.2196 -1.0615 -0.0312 -1.0321 -0.3401 C -1.0228 -0.4383 -1.0375 -0.701 -1.1296 -0.4774 C -1.2393 -0.211 -1.2271 0.0701 -1.2271 0.3522 C -1.2271 0.4359 -1.2232 0.521 -1.2282 0.6047 C -1.2285 0.611 -1.2475 0.7162 -1.2437 0.6512 C -1.225 0.3342 -1.1536 0.0245 -1.1695 -0.2958 C -1.1757 -0.4218 -1.2077 -0.4801 -1.3334 -0.5162 C -1.4217 -0.5416 -1.5639 -0.5748 -1.6391 -0.4996 C -1.8041 -0.3346 -1.7101 0.0591 -1.7421 0.2658 C -1.7545 0.3461 -1.8284 0.7227 -1.9426 0.7553 C -1.9448 0.756 -1.9331 -0.1641 -1.916 -0.2891 C -1.9134 -0.3079 -1.9084 -0.4137 -1.9027 -0.3168 C -1.8904 -0.1073 -1.8721 0.1024 -1.8684 0.3123 C -1.8666 0.4109 -1.8731 0.5071 -1.9138 0.5981 C -1.9208 0.6137 -1.9461 0.6587 -1.9381 0.6435 C -1.8209 0.4187 -1.6434 0.2502 -1.628 -0.0155 C -1.6196 -0.1606 -1.6228 -0.2963 -1.7653 -0.3711 C -1.7997 -0.3891 -2.1243 -0.5003 -2.1253 -0.4929 C -2.1591 -0.2399 -2.3103 -0.033 -2.5385 0.0786 C -2.5492 0.0839 -2.7418 0.142 -2.7301 0.134 C -2.6556 0.0828 -2.5427 0.0667 -2.4576 0.0476 C -2.2715 0.0058 -1.9299 -0.029 -1.9104 -0.2825 C -1.9059 -0.3413 -2.0102 -0.2742 -2.019 -0.2692 C -2.1784 -0.1771 -2.3233 -0.0624 -2.4842 0.0276 C -2.5655 0.0732 -2.8321 0.1428 -2.7389 0.1428 C -2.6716 0.1428 -2.5584 0.078 -2.4997 0.0553 C -2.3851 0.0112 -2.2479 -0.0311 -2.1718 -0.1352 C -2.11 -0.2197 -2.1118 -0.3301 -2.0533 -0.4165 C -1.9699 -0.5397 -1.8375 -0.549 -1.7022 -0.5273 C -1.6385 -0.517 -1.4841 -0.4502 -1.5095 -0.5095 C -1.5566 -0.6195 -1.6924 -0.7271 -1.762 -0.823 C -1.8799 -0.9854 -1.9428 -1.139 -1.9714 -1.3369 C -1.9784 -1.3857 -1.9998 -1.5299 -2.0046 -1.4809 C -2.0252 -1.2694 -2.0268 -1.054 -2.0334 -0.8418 C -2.0397 -0.6359 -2.0445 -0.4298 -2.0445 -0.2238 C -2.0445 -0.1975 -2.0445 -0.1387 -2.0445 -0.2072 C -2.0445 -0.3522 -2.025 -0.51 -1.9193 -0.6203 C -1.8699 -0.6719 -1.9647 -0.4852 -1.9647 -0.4486 C -1.9647 -0.3867 -1.9497 -0.3216 -1.9193 -0.267 C -1.7343 0.0661 -1.3801 0.3161 -1.3312 0.7188 C -1.3249 0.7707 -1.3865 1.0196 -1.3865 0.855 C -1.3865 0.8202 -1.3779 0.7708 -1.3877 0.7365 C -1.3948 0.7114 -1.3584 0.78 -1.3411 0.7996 C -1.1854 0.9777 -1.0001 1.1263 -0.8161 1.2737 C -0.5079 1.5206 -0.0989 1.7578 -0.2402 2.2107 C -0.2893 2.3682 -0.3868 2.5009 -0.4839 2.6316 C -0.5244 2.6862 -0.5855 2.7482 -0.6057 2.8155 C -0.6263 2.8843 -0.5867 2.9587 -0.5714 3.0237 C -0.5646 3.0523 -0.5433 2.9717 -0.5337 2.944 C -0.4822 2.7956 -0.4211 2.6505 -0.3731 2.5009 C -0.2328 2.0636 -0.2332 1.5216 -0.599 1.1951 C -0.771 1.0416 -0.9833 0.9482 -1.1794 0.8318 C -1.2246 0.8049 -1.343 0.7015 -1.2968 0.7265 C -1.1743 0.7929 -1.0687 0.9527 -0.9967 1.0633 C -0.7984 1.3677 -0.7058 1.7096 -0.6389 2.0634 C -0.6042 2.2473 -0.5375 2.4962 -0.5935 2.6859 C -0.5964 2.6958 -0.6703 2.8487 -0.6655 2.8509 C -0.6099 2.8766 -0.4794 2.7824 -0.444 2.7523 C -0.3383 2.6628 -0.2925 2.5321 -0.2767 2.3979 C -0.2114 1.8428 -0.5691 1.4664 -0.9302 1.0998 C -1.0862 0.9415 -1.4944 0.4828 -1.0288 0.4142 C -0.8456 0.3872 -0.4831 0.3627 -0.3698 0.5482 C -0.3356 0.6042 -0.2959 0.7849 -0.218 0.7443 C -0.1508 0.7092 -0.086 0.6474 -0.1028 0.7454 C -0.1784 1.1851 -0.0408 1.6351 0.0843 2.0523 C 0.1285 2.1994 0.1763 2.3562 0.163 2.512 C 0.1608 2.5377 0.0903 2.6684 0.0899 2.6582 C 0.0841 2.5302 0.2259 2.3478 0.2793 2.2417 C 0.3955 2.011 0.4878 1.768 0.4941 1.5063 C 0.4996 1.2786 0.4389 1.0468 0.3479 0.8395 C 0.3377 0.8163 0.2358 0.5723 0.1818 0.5837 C 0.1249 0.5957 0.1317 0.8695 0.1309 0.8971 C 0.1225 1.1656 0.1311 1.434 0.1143 1.7023 C 0.1013 1.9082 0.0764 2.1 -0.0065 2.2894 C -0.0193 2.3186 -0.0544 2.4069 -0.0475 2.3758 C 0.0694 1.8501 0.1931 1.3219 0.2294 0.783 C 0.2371 0.6695 0.1946 0.5526 0.3402 0.5526 C 0.5426 0.5526 0.754 0.5989 0.9472 0.6557 C 1.0977 0.6999 1.2284 0.7838 1.3769 0.8351 C 1.4465 0.8591 1.5178 0.8665 1.5874 0.8872 C 1.6237 0.8979 1.6121 0.9218 1.635 0.9414 C 1.6646 0.9668 1.6935 0.8741 1.7158 0.8672 C 1.7454 0.8581 1.9059 0.8864 1.8997 0.8428 C 1.8971 0.8245 1.8692 0.7776 1.8443 0.7963 C 1.7614 0.8585 1.7763 1.2934 1.7701 1.3568 C 1.7208 1.8617 1.6381 2.3115 1.3459 2.7335 C 1.3412 2.7403 1.3292 2.7617 1.3348 2.7557 C 1.3644 2.7242 1.3858 2.6931 1.4101 2.656 C 1.5196 2.4896 1.5944 2.3102 1.6239 2.1122 C 1.6893 1.6724 1.607 1.2042 1.5475 0.7675 C 1.5137 0.5201 1.7083 1.2447 1.7347 1.493 C 1.7636 1.7651 1.7645 2.0414 1.6749 2.3027 C 1.6558 2.3583 1.6207 2.4809 1.5674 2.5209 C 1.4565 2.6041 1.6884 2.269 1.7214 2.1343 C 1.8035 1.7985 1.7788 1.4455 1.8233 1.1042`)
-let ant_base_paht = new Path2D(`M -2.8 -11.2 c-0.06,0.04 -0.41,0.33 -0.69,0.65c-0.6,0.61 -0.66,0.72 -0.57,1c0,0.06 0.39,0.65 0.81,1.31s0.77,1.23 0.78,1.26a0.29,0.29 0 0 1 -0.07,0.18a3.73,3.73 0 0 0 -0.45,1.08a9.11,9.11 0 0 0 0,2.15a2.09,2.09 0 0 0 0.94,1.25a0.75,0.75 0 0 1 0.24,0.17a0.69,0.69 0 0 1 -0.11,0.18l-0.11,0.16s-0.73,-0.25 -1.62,-0.57l-1.63,-0.6l-1,-1.48c-0.53,-0.81 -1.02,-1.51 -1.08,-1.56a0.5,0.5 0 0 0 -0.85,0.39c0,0.13 0.15,0.38 1.08,1.77c0.62,0.89 1.15,1.65 1.2,1.69s0.86,0.34 1.79,0.68l1.69,0.6l0,0.18a4.17,4.17 0 0 0 -0.05,0.44l0,0.26l-2.14,0a14.88,14.88 0 0 0 -2.23,0.07c-0.17,0.08 -2.29,3.25 -2.34,3.41a0.44,0.44 0 0 0 0,0.26a0.49,0.49 0 0 0 0.73,0.27a13.77,13.77 0 0 0 1.1,-1.54l1,-1.46l4,0l0.06,0.14a3.39,3.39 0 0 0 0.44,0.85c0.14,0.18 0.19,0.29 0.16,0.31l-1.8,0.43c-1.07,0.24 -1.75,0.41 -1.82,0.45s-0.71,1 -1.47,2.21c-1.48,2.28 -1.46,2.26 -1.3,2.54a0.54,0.54 0 0 0 0.69,0.18a19.72,19.72 0 0 0 1.4,-2.07l1.29,-2l1,-0.23c0.56,-0.13 1,-0.23 1,-0.22a1,1 0 0 1 -0.11,0.27c-0.06,0.14 -0.15,0.4 -0.2,0.57a4.71,4.71 0 0 0 1.37,4a2.13,2.13 0 0 0 3.22,0a4.92,4.92 0 0 0 1.44,-3.17a2.72,2.72 0 0 0 -0.2,-1.37a2.05,2.05 0 0 1 -0.13,-0.31s0.48,0.09 1,0.22l1,0.23l1.28,2a24,24 0 0 0 1.39,2.06a0.53,0.53 0 0 0 0.71,-0.16c0.16,-0.28 0.17,-0.26 -1.31,-2.56c-0.75,-1.18 -1.41,-2.17 -1.47,-2.21a16.09,16.09 0 0 0 -1.74,-0.43l-1.73,-0.42s0,-0.11 0.14,-0.27a4.24,4.24 0 0 0 0.5,-1c0,-0.06 0.16,-0.06 2,-0.06l2,0l1,1.47c0.82,1.23 1,1.49 1.12,1.55a0.51,0.51 0 0 0 0.71,-0.53a38.25,38.25 0 0 0 -2.27,-3.42a14.88,14.88 0 0 0 -2.23,-0.07l-2.14,0l0,-0.28c0,-0.16 0,-0.36 -0.05,-0.44l0,-0.16l1.52,-0.57c0.9,-0.32 1.7,-0.61 1.77,-0.65a10.06,10.06 0 0 0 1.23,-1.72c1,-1.51 1.09,-1.65 1.09,-1.81a0.49,0.49 0 0 0 -0.76,-0.42a14.13,14.13 0 0 0 -1.13,1.59l-1,1.5l-1.6,0.57c-0.88,0.32 -1.61,0.58 -1.62,0.57a1.28,1.28 0 0 1 -0.28,-0.35a1,1 0 0 1 0.25,-0.16a2.19,2.19 0 0 0 1,-1.28a10.74,10.74 0 0 0 0,-2.12a3.15,3.15 0 0 0 -0.41,-1a1,1 0 0 1 -0.12,-0.26s0.28,-0.58 0.73,-1.26l0.81,-1.29a0.47,0.47 0 0 0 0,-0.43a9.12,9.12 0 0 0 -1.36,-1.28a0.48,0.48 0 0 0 -0.5,0.16a0.39,0.39 0 0 0 -0.13,0.34l0,0.23l0.44,0.44l0.43,0.45l-0.59,0.94c-0.33,0.51 -0.61,0.95 -0.63,1s-0.13,0 -0.26,-0.09a3.06,3.06 0 0 0 -2.84,0a1,1 0 0 1 -0.3,0.12s-0.3,-0.46 -0.63,-1l-0.61,-0.94l0.44,-0.45l0.43,-0.47l0,-0.21a0.41,0.41 0 0 0 -0.15,-0.35a0.48,0.48 0 0 0 -0.59,-0.1z`)
-
-ctx.strokeStyle  = "black";
-ctx.lineWidth = 1 
-ctx.fillStyle = "black";
-function anotherFigure(x,y,angle_radian,type){
-    // let angleInRadians = angle_radian * Math.PI/180    // ctx.translate(x_center,y_center)
-    ctx.translate(x, y);
-    ctx.rotate(angle_radian)
-    ctx.fill(type)
-    // ctx.lineWidth = 0.1
-    // ctx.scale(5,5)
-    // ctx.stroke(type)
-    ctx.resetTransform();
-}
-function drawFullCadrAnt(obj_ant){
-    let main_obj_ant = obj_ant.Ant
-    let angle_ant = main_obj_ant.angle
-    for(let i = 0;i<main_obj_ant.pos.length;i++){
-        let x = main_obj_ant.pos[i][0]
-        let y = main_obj_ant.pos[i][1]
-        let angle_radian = angle_ant[i]
-        anotherFigure(x,y,angle_radian,ant_base_paht)
-    }
-
-
-
-
-
-    
-    // let array_coordinates = obj_ant.coordinates
-
-    // for(let i = 0;i<array_coordinates.length;i++){
-    //     let x =array_coordinates[i]["x"]
-    //     let y =array_coordinates[i]["y"]
-    //     let angle =array_coordinates[i]["angle"]
-    //     anotherFigure(x,y,angle,ant_base_paht)
-    // }
-}
-function drawFullCadrMem(obj_ant){
-    // let array_coordinates = obj_ant.coordinates
-    // for(let i = 0;i<array_coordinates.length;i++){
-    //     let x =array_coordinates[i]["x"]
-    //     let y =array_coordinates[i]["y"]
-    //     let angle =array_coordinates[i]["angle"]
-    //     anotherFigure(x,y,angle,among_us)
-    // }
-    let main_obj_ant = obj_ant.Ant
-    let angle_ant = main_obj_ant.angle
-    for(let i = 0;i<main_obj_ant.pos.length;i++){
-        let x = main_obj_ant.pos[i][0]
-        let y = main_obj_ant.pos[i][1]
-        let angle_radian = angle_ant[i]
-        anotherFigure(x,y,angle_radian,among_us_origin)
-    }
-}
-function drawPoint(obj_ant){
-    let array_coordinates = obj_ant.coordinates
-    for(let i = 0;i<array_coordinates.length;i++){
-        let x =array_coordinates[i]["x"]
-        let y =array_coordinates[i]["y"]
-        let angle =array_coordinates[i]["angle"]
-        anotherFigure(x,y,angle,ant_base_paht)
-    }
-}
-
-function clearCanvas() {
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-}
-export {
-    anotherFigure,
-    drawFullCadrAnt,
-    clearCanvas,
-    drawFullCadrMem,
-    drawPoint
-}
