+++ Ready_ant_server/static/html/index.html
@@ -143,8 +143,9 @@
     </footer>
     <script type="module" src="../script/client.js"></script>
 
-    <script type="module" src="../draw_cansvas/setting_canvas.js"></script>
-    <script type="module" src="../draw_cansvas/draw.js"></script>
+    <script type="module" src="../script/draw_cansvas/setting_canvas.js"></script>
+    <script type="module" src="../script/draw_cansvas/draw.js"></script>
+    <script type="module" src="../script/draw_cansvas/draw.js"></script>
 </body>
 
 </html>
