+++ Ready_ant_server/static/script/client.js
@@ -1,5 +1,5 @@
 // import{drawAnt} from "./draw_cansvas/draw.js"
-import{drawAnt} from "../draw_cansvas/draw.js"
+import{drawAnt} from "../script/draw_cansvas/draw.js"
 
 const channel = geckos({ port: 3000 })
